### 2025-01-21 #7705 @editorialbot
[PRE REVIEW]: AquaFetch: A Unified Python Interface for Water Resource Dataset Acquisition and Harmonization
**Submitting author:** @AtrCheema (<a href="http://orcid.org/0000-0002-0031-745X">Ather Abbas</a>)
**Repository:** https://github.com/hyex-research/AquaFetch
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
