### 2025-01-22 852b7e2
added paper badge and updated bib
