+++ paper/paper.bib
@@ -4,7 +4,8 @@
   booktitle = "Proceedings of the 9th Python in Science Conference",
   pages     = "51--56",
   year      = 2010,
-  editor    = "Stefan van der Walt and Jarrod Millman"
+  editor    = "Stefan van der Walt and Jarrod Millman",
+DOI = {10.25080/Majora-92bf1922-00a}
 }
 
 
@@ -133,8 +134,8 @@
 
 
 @article{weathercan,
-  doi = {10.21105/joss.0057},
-  url = {https://doi.org/10.21105/joss.0057},
+  doi = {10.21105/joss.00571},
+  url = {https://doi.org/10.21105/joss.00571},
   year = {2018},
   publisher = {The Open Journal},
   volume = {3},
@@ -233,7 +234,8 @@ DOI = {10.5194/essd-13-4529-2021}
   number={1},
   pages={879},
   year={2024},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+DOI = {10.1038/s41597-024-03706-1}
 }
 
 
@@ -261,7 +263,8 @@ DOI = {10.5194/essd-16-1559-2024}
   number={1},
   pages={243},
   year={2020},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+DOI = {10.1038/s41597-020-00583-2}
 }
 
 
@@ -274,7 +277,8 @@ DOI = {10.5194/essd-16-1559-2024}
   number={1},
   pages={737},
   year={2024},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+DOI = {10.1038/s41597-024-03594-5}
 }
 
 
@@ -324,9 +328,9 @@ DOI = {10.5194/essd-2024-292}
 
 
 @misc{requests,
-  author = {Aleksey Rogozhnikov},
-  title = {Einops},
-  howpublished = {\url{https://github.com/arogozhnikov/einops}},
+  author = { Kenneth Reitz},
+  title = {requests: A simple, yet elegant, HTTP library},
+  howpublished = {\url{https://github.com/psf/requests}},
   year = {2018},
   note = {Accessed on 16 Jan 2024}
 }
@@ -371,7 +375,6 @@ url = {https://github.com/Unidata/netcdf-c}
     https://aip.scitation.org/doi/pdf/10.1109/MCSE.2007.55
   }
 }
-
 
 @Article{camels_gb,
 AUTHOR = {Coxon, G. and Addor, N. and Bloomfield, J. P. and Freer, J. and Fry, M. and Hannaford, J. and Howden, N. J. K. and Lane, R. and Lewis, M. and Robinson, E. L. and Wagener, T. and Woods, R.},
