### 2025-01-22 50dcec6
removed minor bugs for estreams and docs improvement
