### 2025-01-26 47f3d19
removed some bugs in lamah
