+++ aqua_fetch/rr/_lamah.py
@@ -17,6 +17,9 @@ from .utils import _RainfallRunoff, _handle_dynamic
 from ._map import (
     observed_streamflow_cms,
     min_air_temp,
+    min_air_temp_with_specifier,
+    max_air_temp_with_specifier,
+    mean_air_temp_with_specifier,
     max_air_temp,
     mean_air_temp,
     total_precipitation,
@@ -32,6 +35,7 @@ from ._map import (
     min_dewpoint_temperature_at_2m,
     mean_dewpoint_temperature_at_2m,
     mean_air_pressure,
+    total_potential_evapotranspiration,
 )
 
 from ._map import (
@@ -104,11 +108,11 @@ class LamaHCE(_RainfallRunoff):
         >>> dataset = LamaHCE(timestep='H', data_type='total_upstrm')
         >>> len(dataset.stations()), len(dataset.static_features), len(dataset.dynamic_features)
         (859, 80, 17)
-        >>> dataset.fetch_dynamic_features('1', features = ['obs_q_cms'])
+        >>> dataset.fetch_dynamic_features('1', features = ['q_cms_obs'])
         """
 
-        assert timestep in self.time_steps, f"invalid timestep {timestep} given"
-        assert data_type in self._data_types, f"invalid data_type {data_type} given."
+        assert timestep in self.time_steps, f"invalid timestep '{timestep}' given, choose from {self.time_steps}"
+        assert data_type in self._data_types, f"invalid data_type '{data_type}' given, choose from {self._data_types}"
 
         self.timestep = timestep
         self.data_type = data_type
@@ -142,11 +146,10 @@ class LamaHCE(_RainfallRunoff):
     @property
     def static_map(self) -> Dict[str, str]:
         return {
-                'area_calc': catchment_area(),
+                'area_calc': catchment_area(), # todo : difference between area_calc and area_gov?
                 'lat': gauge_latitude(),
                 'slope_mean': slope('mkm-1'),
                 'lon': gauge_longitude(),
-
         }
 
     @property
@@ -207,6 +210,9 @@ class LamaHCE(_RainfallRunoff):
         # since data is very large, saving all the data in one file
         # consumes a lot of memory, which is impractical for most of the personal
         # computers! Therefore, saving each feature separately
+
+        # todo: if we are only interested in one dynamic feature say 'o_cms_obs', 
+        # then why do we need to save all the dynamic features in the netcdf file?
 
         fdir = os.path.join(self.path, fdir)
         if not os.path.exists(fdir):
@@ -337,13 +343,13 @@ class LamaHCE(_RainfallRunoff):
 
         Examples
         --------
-            >>> from aqua_fetch import CAMELS_AUS
-            >>> dataset = CAMELS_AUS()
-            ... # find out station ids
-            >>> dataset.stations()
-            ... # get data of selected stations
-            >>> dataset.fetch_stations_features(['912101A', '912105A', '915011A'],
-            ...  as_dataframe=True)
+        >>> from aqua_fetch import CAMELS_AUS
+        >>> dataset = CAMELS_AUS()
+        ... # find out station ids
+        >>> dataset.stations()
+        ... # get data of selected stations
+        >>> dataset.fetch_stations_features(['912101A', '912105A', '915011A'],
+        ...  as_dataframe=True)
         """
         st, en = self._check_length(st, en)
 
@@ -385,16 +391,16 @@ class LamaHCE(_RainfallRunoff):
 
     @property
     def _q_name(self) -> str:
-        return 'obs_q_cms'
+        return observed_streamflow_cms()
 
-    @property
-    def _area_name(self) -> str:
-        # todo : difference between area_calc and area_gov?
-        return 'area_calc'
+    # @property
+    # def _area_name(self) -> str:
+    #     
+    #     return catchment_area()
 
-    @property
[…]
