+++ aqua_fetch/rr/_camels.py
@@ -73,7 +73,7 @@ class CAMELS_US(_RainfallRunoff):
     --------
     >>> from aqua_fetch import CAMELS_US
     >>> dataset = CAMELS_US()
-    >>> df = dataset.fetch(stations=1, as_dataframe=True)
+    >>> _, df = dataset.fetch(stations=1, as_dataframe=True)
     >>> df = df.unstack() # the returned dataframe is a multi-indexed dataframe so we have to unstack it
     >>> df.shape
     (12784, 8)
@@ -82,33 +82,32 @@ class CAMELS_US(_RainfallRunoff):
     >>> len(stns)
     671
     # we can get data of 10% catchments as below
-    >>> data = dataset.fetch(0.1, as_dataframe=True)
+    >>> _, data = dataset.fetch(0.1, as_dataframe=True)
     >>> data.shape
     (460488, 51)
     # the data is multi-index with ``time`` and ``dynamic_features`` as indices
     >>> data.index.names == ['time', 'dynamic_features']
      True
     # get data by station id
-    >>> df = dataset.fetch(stations='11478500', as_dataframe=True).unstack()
+    >>> _, df = dataset.fetch(stations='11478500', as_dataframe=True).unstack()
     >>> df.shape
     (12784, 8)
     # get names of available dynamic features
     >>> dataset.dynamic_features
     # get only selected dynamic features
-    >>> df = dataset.fetch(1, as_dataframe=True,
+    >>> _, df = dataset.fetch(1, as_dataframe=True,
     ... dynamic_features=['prcp(mm/day)', 'srad(W/m2)', 'tmax(C)', 'tmin(C)', 'Flow']).unstack()
     >>> df.shape
     (12784, 5)
     # get names of available static features
     >>> dataset.static_features
     # get data of 10 random stations
-    >>> df = dataset.fetch(10, as_dataframe=True)
+    >>> _, df = dataset.fetch(10, as_dataframe=True)
     >>> df.shape
     (102272, 10)  # remember this is multi-indexed DataFrame
-    # when we get both static and dynamic data, the returned data is a dictionary
-    # with ``static`` and ``dyanic`` keys.
-    >>> data = dataset.fetch(stations='11478500', static_features="all", as_dataframe=True)
-    >>> data['static'].shape, data['dynamic'].shape
+    # If we want to get both static and dynamic data 
+    >>> static, dynamic = dataset.fetch(stations='11478500', static_features="all", as_dataframe=True)
+    >>> static.shape, dynamic.shape
     ((1, 59), (102272, 1))
 
     """
@@ -170,6 +169,7 @@ class CAMELS_US(_RainfallRunoff):
         self.attr_dir = os.path.join(self.path, f'catchment_attrs{SEP}camels_attributes_v2.0')
         self.dataset_dir = os.path.join(self.path, f'CAMELS_US{SEP}basin_dataset_public_v1p2')
 
+        self._static_features = self._read_static().columns.tolist()
         self._maybe_to_netcdf('camels_us_dyn')
 
     @property
@@ -209,35 +209,12 @@ class CAMELS_US(_RainfallRunoff):
         return "20141231"
 
     @property
-    def static_features(self):
-        static_fpath = os.path.join(self.path, 'static_features.csv')
-        if not os.path.exists(static_fpath):
-            files = glob.glob(f"{os.path.join(self.path, 'catchment_attrs', 'camels_attributes_v2.0')}/*.txt")
-            cols = []
-            for f in files:
-                _df = pd.read_csv(f, sep=';', index_col='gauge_id', nrows=1)
-                cols += list(_df.columns)
-        else:
-            df = pd.read_csv(static_fpath, index_col='gauge_id', nrows=1)
[…]
