+++ tests/rr/test_bull.py
@@ -1,10 +1,10 @@
 
 import os
-import site   # so that water_quality directory is in path
+import site   # so that aqua_fetch directory is in path
 import logging
 
 # add the parent directory in the path
-wd_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+wd_dir = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 site.addsitedir(wd_dir)
 
 gscad_path = '/mnt/datawaha/hyex/atr/gscad_database/raw'
@@ -16,11 +16,7 @@ logger = logging.getLogger(__name__)
 
 from aqua_fetch import Bull
 
-from utils import (
-    test_dataset,
-    test_dynamic_data,
-    test_attributes
-    )
+from utils import test_dataset
 
 
 dataset = Bull(path=gscad_path, verbosity=3)
