+++ aqua_fetch/rr/_gsha.py
@@ -1123,7 +1123,7 @@ class _GSHA(_RainfallRunoff):
             en=None,
             as_dataframe: bool = False,
             **kwargs
-              ) -> Tuple[pd.DataFrame, Union[pd.DataFrame: xr.Dataset]]:
+              ) -> Tuple[pd.DataFrame, Union[pd.DataFrame, xr.Dataset]]:
         """
         returns features of multiple stations
 
