### 2025-01-29 d4b32ee
improved npctr dataset and docs
