+++ aqua_fetch/__init__.py
@@ -264,4 +264,4 @@ def load_nasdaq(inputs: Union[str, list, None] = None, target: str = 'NDX'):
     return df[inputs + target]
 
 
-__version__ = "1.0.0rc1"
+__version__ = "1.0.0rc2"
