+++ aqua_fetch/rr/_npctr.py
@@ -2,20 +2,55 @@
 __all__ = ["NPCTRCatchments"]
 
 import os
-from typing import Union, List
+from typing import Union, List, Dict
 
+import numpy as np
 import pandas as pd
 
-from .._datasets import Datasets
+from .._backend import shapefile
 from ..utils import check_attributes
+from .utils import _RainfallRunoff
+from ._map import(
+    observed_streamflow_cms,
+    mean_air_temp,
+    mean_rel_hum,
+    mean_windspeed,
+    solar_radiation,
+    total_precipitation,
+    snow_depth,
+)
 
+from ._map import (
+    catchment_area,
+    gauge_longitude,
+    gauge_latitude,
+    slope,
+    catchment_elevation_meters,
+    max_catchment_elevation_meters,
+)
 
-class NPCTRCatchments(Datasets):
+# todo : what is difference between SSN and WSN stations?
+
+class NPCTRCatchments(_RainfallRunoff):
     """
     High-resolution streamflow and weather data (2013–2019) for seven small coastal
     watersheds in the northeast Pacific coastal temperate rainforest, Canada following
-    `Korver et al., 2022 <https://doi.org/10.5194/essd-14-4231-2022>`_
+    `Korver et al., 2022 <https://doi.org/10.5194/essd-14-4231-2022>`_ . The data
+    include 8 dynamic features at hourly and 5 min timestep and 14 static features.
+    The dynamic features include streamflow, precipitation, temperature, relative humidity,
+    wind speed, wind direction, and solar radiation.
 
+    Examples
+    --------
+    >>> from aqua_fetch import NPCTRCatchments
+    >>> ds = NPCTRCatchments()
+    >>> ds.stations
+    ['626', '693', '703', '708', '819', '844', '1015']
+    >>> len(ds.static_features)
+    12
+    >>> area = ds.area()
+    >>> area.shape
+    (7,)
     """
     url = {
         "2013-2019_Discharge1015_5min.csv":
@@ -62,94 +97,459 @@ class NPCTRCatchments(Datasets):
             "https://media.githubusercontent.com/media/HakaiInstitute/essd2021-hydromet-datapackage/main/2013-2019_WindSpd_Hourly.csv",
         "Data-Dictonary.csv":
             "https://media.githubusercontent.com/media/HakaiInstitute/essd2021-hydromet-datapackage/main/Data-Dictonary.csv",
+        "Watersheds_lidar_derived.zip":
+            "https://drive.google.com/file/d/1TzrDVZszFI0Y44WbRJCAuReYUpR8_wcx/view?usp=drive_link",
+        "Focal_watersheds_lidar_derived.zip":
+            "https://drive.google.com/file/d/1W6qP_9wqintvWlsVrR6FBByaelmJAIBp/view?usp=drive_link",       
+        "watershedsmedata.pdf":
+            "https://drive.google.com/file/d/1knjwA7EuI5L8rlzMkdsUvfN5CyxwMfL5/view?usp=drive_link",
     }
 
-    def __init__(self, path=None, **kwargs):
+    def __init__(self, path=None,
+                 timestep:str = "Hourly",
+                 **kwargs):
         super().__init__(path=path, **kwargs)
-        self.ds_dir = path
+
         self._download()
 
-    @property
+        self.timestep = _verify_timestep(timestep)
+
+        self.boundary_file = os.path.join(
+            self.path, 
+            "Focal_watersheds_lidar_derived",
+            "focal_watersheds_lidar_derived.shp"
+            )
+        
+        self._create_boundary_id_map(self.boundary_file, 0)
+
+        self._static_features = self._get_static().columns.tolist()
+
     def stations(self)->List[str]:
         return ["626", "693", "703", "708", "819", "844", "1015"]
 
-    def fetch_wind_speed(
-            self,
[…]
