+++ docs/requirements.txt
@@ -10,6 +10,7 @@ nbsphinx
 ipykernel
 
 git+https://github.com/AtrCheema/tabulight.git  # for examples
+basemap    # for example scripts
 
 openpyxl
 seaborn
