### 2025-01-29 6fe2c3d
improved docs and npctr
