+++ aqua_fetch/rr/_npctr.py
@@ -29,7 +29,7 @@ from ._map import (
     max_catchment_elevation_meters,
 )
 
-# todo : what is difference between SSN and WSN stations?
+# SSN : Stream Sensor Node and WSN : Weather Sensor Node
 
 class NPCTRCatchments(_RainfallRunoff):
     """
@@ -51,6 +51,9 @@ class NPCTRCatchments(_RainfallRunoff):
     >>> area = ds.area()
     >>> area.shape
     (7,)
+    >>> coords = ds.stn_coords()
+    >>> coords.shape
+    (7, 2)
     """
     url = {
         "2013-2019_Discharge1015_5min.csv":
@@ -139,7 +142,51 @@ class NPCTRCatchments(_RainfallRunoff):
     def q_attributes(self):
         return ["Qrate", "Qrate_min", "Qrate_max", "Qvol", "Qvol_min", "Qvol_max",
                 "Qmm", "Qmm_min", "Qmm_max"]
-    
+
+    def stn_coords(self, stations = 'all', sensor='SSN')->pd.DataFrame:
+        """
+        By default uses coordinate information of Stream Sensor Nodes, assuming that
+        stream sensors would be closer to the stream gauge. The values are taken
+        from `Table A1 of paper <https://essd.copernicus.org/articles/14/4231/2022/essd-14-4231-2022.html#&gid=1&pid=1>`_
+        """        
+        coords = self.all_stn_coords()
+
+        if sensor == 'SSN':
+            coords.index = coords.index.astype(str).str.strip('SSN')
+        elif sensor == 'WSN':
+            coords.index = coords.index.astype(str).str.strip('WSN')
+
+        stations = check_attributes(stations, self.stations(), 'stations')
+        
+        return coords.loc[stations, :]
+
+    def all_stn_coords(self)->pd.DataFrame:
+        """
+        Using coordinate information of Stream Sensor Nodes, assuming that
+        stream sensors would be closer to the stream gauge. The values are taken
+        from `Table A1 of paper <https://essd.copernicus.org/articles/14/4231/2022/essd-14-4231-2022.html#&gid=1&pid=1>`_
+        """
+        stations = {
+            "RefStn": {gauge_latitude: 51.6520, gauge_longitude(): -128.1287},
+            "SSN626": {gauge_latitude: 51.6408, gauge_longitude(): -128.1219},
+            "WSN626": {gauge_latitude: 51.6262, gauge_longitude(): -128.1018},
+            "SSN693": {gauge_latitude: 51.6442, gauge_longitude(): -127.9978},
+            "WSN693_703": {gauge_latitude: 51.6106, gauge_longitude(): -127.9871},
+            "SSN703": {gauge_latitude: 51.6166, gauge_longitude(): -128.0257},
+            "WSN703": {gauge_latitude: 51.6433, gauge_longitude(): -128.0228},
+            "WSN703_708": {gauge_latitude: 51.6222, gauge_longitude(): -128.0507},
+            "SSN708": {gauge_latitude: 51.6486, gauge_longitude(): -128.0684},
+            "SSN819": {gauge_latitude: 51.6619, gauge_longitude(): -128.0419},
+            "WSN819_1015": {gauge_latitude: 51.6827, gauge_longitude(): -128.0433},
+            "SSN844": {gauge_latitude: 51.6608, gauge_longitude(): -128.0025},
+            "WSN844": {gauge_latitude: 51.6614, gauge_longitude(): -127.9975},
+            "SSN1015": {gauge_latitude: 51.6906, gauge_longitude(): -128.0653},
+            "East Buxton": {gauge_latitude: 51.5899, gauge_longitude(): -128.9752},
+            "Hecate": {gauge_latitude: 51.6826, gauge_longitude():-128.0228}
+        }
+
[…]
