+++ aqua_fetch/rr/__init__.py
@@ -53,6 +53,7 @@ from ._npctr import NPCTRCatchments
 from .mtropics import MtropicsLaos
 from .mtropics import MtropcsThailand
 from .mtropics import MtropicsVietnam
+from ._misc import DraixBleone
 
 
 DATASETS = {
