+++ aqua_fetch/rr/_gsha.py
@@ -1123,7 +1123,7 @@ class _GSHA(_RainfallRunoff):
             en=None,
             as_dataframe: bool = False,
             **kwargs
-              ) -> Tuple[pd.DataFrame, Union[pd.DataFrame, xr.Dataset]]:
+              ) -> Tuple[pd.DataFrame, Union[pd.DataFrame, "Dataset"]]:
         """
         returns features of multiple stations
 
@@ -1138,10 +1138,10 @@ class _GSHA(_RainfallRunoff):
         -------
         tuple
             A tuple of static and dynamic features. Static features are always
-            returned as pandas DataFrame with shape (stations, staticfeatures).
+            returned as :obj:`pandas.DataFrame` with shape (stations, staticfeatures).
             The index of static features is the station/gauge ids while the columns 
             are the static features. Dynamic features are returned as either
-            xarray Dataset or pandas DataFrame depending upon whether `as_dataframe`
+            xarray Dataset or :obj:`pandas.DataFrame` depending upon whether `as_dataframe`
             is True or False and whether the xarray module is installed or not.
             If dynamic features are xarray Dataset, then it consists of `data_vars`
             equal to the number of stations and `time` adn `dynamic_features` as
