+++ aqua_fetch/rr/_hysets.py
@@ -456,7 +456,7 @@ class HYSETS(_RainfallRunoff):
             en=None,
             as_dataframe: bool = False,
             **kwargs
-              ) -> Tuple[pd.DataFrame, Union[pd.DataFrame, xr.Dataset]]:
+              ) -> Tuple[pd.DataFrame, Union[pd.DataFrame, "Dataset"]]:
         """returns features of multiple stations
         Examples
         --------
