+++ aqua_fetch/utils.py
@@ -930,7 +930,7 @@ def get_version_info()->dict:
     versions = {
         'numpy': np.__version__,
         'pandas': pd.__version__,
-        'water_quality': __version__,
+        'aqua_fetch': __version__,
         'python': sys.version,
         'os': os.name
     }
