+++ setup.py
@@ -32,7 +32,7 @@ extra_requires = [
 
 # for reading data
 'netCDF4',
-'xarray',
+'xarray<2025.1.0',  # xarray 2025.1.1 is causing to_netcdf error
 
 # todo : following libraries are required by read_html
 #lxml for reading html
