+++ aqua_fetch/rr/__init__.py
@@ -95,66 +95,7 @@ DATASETS = {
     'Portugal': Portugal,
 }
 
-"""
-    .. list-table:: Naming Convention for dynamic features
-       :widths: 20 30
-       :header-rows: 1
 
-       * - Feature Name
-         - Description
-       * - obs_q_cms
-         - observed streamflow in cms
-       * - obs_q_mmd
-         - observed streamflow in mm/day
-       * - pcp_mm
-         - precipitation in mm
-       * - max_temp_C
-         - maximum air temperature in degree celcius
-       * - min_temp_C
-         - minimum air temperature in degree celcius
-       * - mean_temp_C
-         - mean temperature in degree celcius
-       * - method_pet_mm
-         - potential evapotranspiration in mm
-       * - et_mm
-         - evapotranspiration in mm
-       * - rh_%
-         - relative humidity in percentage
-       * - min_rh_%
-         - minimum relative humidity
-       * - max_rh_%
-         - maximum relative humidity
-       * - swe_mm
-         - snow water equivalent
-       * - solrad_wm2
-         - solar radiation watt per meter square
-       * - windspeed_ms
-         - wind speed in meter per second
-       * - sim_q_cms
-         - simulated streamflow in cms
-       * - dwn_lw_rad_wm2
-         - downward long wave radiation in watt per meter square
-       * - dwn_sw_rad_wm2
-         - downward short wave radiation in watt per meter square
-       * - airpres_hpa
-         - Mean air pressure at sea level in hectopascal
-
-
-    .. list-table:: Naming Convention for static features
-       :widths: 20 30
-       :header-rows: 1
-
-       * - Feature Name
-         - Description
-       * - guage_lat
-         - Latitude of the guage station
-       * - guage_long
-         - Longitude of the guage station
-       * - area_km2
-         - catchment area in km2
-       * - mean_elev
-         - mean elevation in meters
-"""
 class RainfallRunoff(object):
     """
     This  class provides access to all the rainfall-runoff
@@ -376,7 +317,7 @@ class RainfallRunoff(object):
         Returns
         -------
         pd.DataFrame
-            a pandas dataframe
+            a pandas :obj:`pandas.DataFrame`
 
         Examples
         --------
@@ -437,7 +378,7 @@ class RainfallRunoff(object):
         stations :
             It can have following values:
 
-                - int : number of (randomly selected) stations to fetch
+                - :obj:`int` : number of (randomly selected) stations to fetch
                 - float : fraction of (randomly selected) stations to fetch
                 - str : name/id of station to fetch. However, if ``all`` is
                   provided, then all stations will be fetched.
@@ -448,14 +389,14 @@ class RainfallRunoff(object):
                 - str : name of dynamic feature to fetch. If ``all`` is
                   provided, then all dynamic features will be fetched.
                 - list : list of dynamic features to fetch.
-                - None : No dynamic feature will be fetched.
+                - None : No dynamic feature will be fetched. The second returned value will be None.
         static_features : (default=None)
             It can have following values:
 
                 - str : name of static feature to fetch. If ``all`` is
                   provided, then all static features will be fetched.
                 - list : list of static features to fetch.
-                - None : No static feature will be fetched.
+                - None : No static feature will be fetched. The first returned value will be None.
         st :
             starting date of data to be returned. If None, the data will be
             returned from where it is available.
@@ -464,7 +405,7 @@ class RainfallRunoff(object):
             returned till the date data is available.
         as_dataframe :
             whether to return dynamic attributes as pandas
-            dataframe or as xarray dataset.
+            dataframe or as :obj:`xarray.Dataset`.
         kwargs :
             keyword arguments
 
[…]
