+++ aqua_fetch/rr/__init__.py
@@ -9,7 +9,7 @@ import os
 from typing import Union, List
 
 import pandas as pd
-from .._backend import plt, plt_Axes, xarray as xr
+from .._backend import plt, plt_Axes
 
 from .utils import _RainfallRunoff
 from ._camels import CAMELS_AUS
@@ -104,7 +104,7 @@ class RainfallRunoff(object):
 
     Examples
     --------
-    >>> from water_datasets import RainfallRunoff
+    >>> from aqua_fetch import RainfallRunoff
     >>> dataset = RainfallRunoff('CAMELS_AUS')  # instead of CAMELS_AUS, you can provide any other dataset name
     >>> _, df = dataset.fetch(stations=1, as_dataframe=True)
     >>> df = df.unstack() # the returned dataframe is a multi-indexed dataframe so we have to unstack it
@@ -215,20 +215,20 @@ class RainfallRunoff(object):
             from this path. If provided and the path/dataset does not exist,
             then the data will be downloaded at this path. If not provided,
             then the data will be downloaded in the default path which is
-            ``.../water-datasts/data/``.
+            ``.../aqua_fetch/data/``.
         overwrite : bool
             If the data is already downloaded then you can set it to True,
             to make a fresh download.
         to_netcdf : bool
             whether to convert all the data into one netcdf file or not.
             This will fasten repeated calls to fetch etc but will
-            require netcdf5 package as well as xarray.
+            require netcdf5 package as well as :obj:`xarray`.
         verbosity : int
             0: no message will be printed
         kwargs :
             additional keyword arguments for the underlying dataset class
-            For example ``version`` for :py:class:`water_quality.rr.CAMELS_AUS` or ``timestep`` for
-            :py:class:`water_quality.rr.LamaHCE` dataset or ``met_src`` for ``CAMELS_BR``
+            For example ``version`` for :py:class:`aqua_fetch.rr.CAMELS_AUS` or ``timestep`` for
+            :py:class:`aqua_fetch.rr.LamaHCE` dataset or ``met_src`` for ``CAMELS_BR``
         """
 
         if dataset not in DATASETS:
@@ -269,7 +269,7 @@ class RainfallRunoff(object):
     def path(self) -> str:
         """
         returns path where the data is stored. The default path is
-        ~../water_quality/data
+        ~../aqua_fetch/data
         """
         return self.dataset.path
 
@@ -280,7 +280,7 @@ class RainfallRunoff(object):
 
         Examples
         --------
-        >>> from water_datasets import RainfallRunoff
+        >>> from aqua_fetch import RainfallRunoff
         >>> dataset = RainfallRunoff('CAMELS_AUS')
         >>> dataset.static_features
         """
@@ -293,7 +293,7 @@ class RainfallRunoff(object):
 
         Examples
         --------
-        >>> from water_datasets import RainfallRunoff
+        >>> from aqua_fetch import RainfallRunoff
         >>> dataset = RainfallRunoff('CAMELS_AUS')
         >>> dataset.dynamic_features
         """
@@ -321,7 +321,7 @@ class RainfallRunoff(object):
 
         Examples
         --------
-        >>> from water_datasets import RainfallRunoff
+        >>> from aqua_fetch import RainfallRunoff
         >>> camels = RainfallRunoff('CAMELS_AUS')
         >>> camels.fetch_static_features('224214A')
         >>> camels.static_features
@@ -352,7 +352,7 @@ class RainfallRunoff(object):
 
         Examples
         ---------
-        >>> from water_datasets import RainfallRunoff
+        >>> from aqua_fetch import RainfallRunoff
         >>> dataset = RainfallRunoff('CAMELS_CH')
         >>> dataset.area()  # returns area of all stations
         >>> dataset.area('2004')  # returns area of station whose id is 2004
[…]
