+++ tests/rr/test_lamah.py
@@ -1,6 +1,6 @@
 
 import os
-import site   # so that water_datasets directory is in path
+import site   # so that aqua_fetch directory is in path
 import logging
 
 # add the parent directory in the path
