+++ tests/rr/test_grdccaravan.py
@@ -1,6 +1,6 @@
 
 import os
-import site   # so that water_quality directory is in path
+import site   # so that aqua_fetch directory is in path
 import logging
 
 # add the parent directory in the path
