### 2025-02-04 3a2aa43
always use station and stations instead of stn_id etc
