+++ aqua_fetch/rr/mtropics.py
@@ -200,7 +200,7 @@ class MtropicsLaos(Datasets):
 
         Returns
         -------
-            a pandas dataframe
+            a :obj:`pandas.DataFrame`
 
         Examples
         --------
@@ -275,7 +275,7 @@ class MtropicsLaos(Datasets):
 
         Returns
         -------
-            a pandas dataframe consisting of features as columns.
+            a :obj:`pandas.DataFrame` consisting of features as columns.
 
         .. _Ribolzi:
             https://dataverse.ird.fr/dataset.xhtml?persistentId=doi:10.23708/EWOYNK
@@ -418,7 +418,7 @@ class MtropicsLaos(Datasets):
                 frequency at which the data is to be fetched.
         Returns
         -------
-            a pandas dataframe consisting of 4 columns
+            a :obj:`pandas.DataFrame` consisting of 4 columns
 
         .. [1]:
             https://doi.org/10.1038/s41598-017-04385-2
@@ -511,7 +511,7 @@ class MtropicsLaos(Datasets):
 
         Returns
         -------
-            pandas dataframe of precipitation data
+            :obj:`pandas.DataFrame` of precipitation data
 
         .. _precipitation:
             https://doi.org/10.1038/s41598-017-04385-2
