### 2025-02-04 87aa08f
refactored rr code to make it DRY
