+++ aqua_fetch/rr/_simbi.py
@@ -526,7 +526,7 @@ class Simbi(_RainfallRunoff):
         'pcp': 'pcp_mm',
         }
 
-    def _read_dynamic_from_csv(
+    def _read_dynamic(
         self,
         stations,
         dynamic_features,

[…]
