### 2025-02-06 #7705 @kthyng
Please reduce to length of your paper to up to about 1000 words. Paper format information here: https://joss.readthedocs.io/en/latest/paper.html. You can check paper length (and a few other items) with the command `@editorialbot check repository` though note it needs to be the first item in a comment to run correctly. Currently the paper has 3284 words. Also you might be missing the DOIs noted above.
