### 2025-02-16 84deece
set default values for all input args for LamaHCE
