### 2025-02-17 889281a
removed as_ts argument for static features
