+++ aqua_fetch/rr/_estreams.py
@@ -496,7 +496,6 @@ class _EStreams(_RainfallRunoff):
             st=None,
             en=None,
             as_dataframe=False,
-            as_ts=False
     ):
         """Fetches dynamic features of station."""
         st, en = self._check_length(st, en)
@@ -546,7 +545,6 @@ class _EStreams(_RainfallRunoff):
             static_features: Union[str, list] = 'all',
             st=None,
             en=None,
-            as_ts=False
     ) -> pd.DataFrame:
         """Fetches static features of station."""
         if self.verbosity > 1:
@@ -621,7 +619,6 @@ class _EStreams(_RainfallRunoff):
             static_features: Union[str, List[str]] = "all",
             st=None,
             en=None,
-            as_ts=False
     ) -> pd.DataFrame:
         """
         returns static atttributes of one or multiple stations
@@ -635,7 +632,6 @@ class _EStreams(_RainfallRunoff):
                 static features are returned.
             st :
             en :
-            as_ts :
 
         Examples
         ---------
@@ -660,7 +656,7 @@ class _EStreams(_RainfallRunoff):
         >>> static_data.shape
            (12004, 2)
         """
-        return self._fetch_static_features(stations, static_features, st, en, as_ts)
+        return self._fetch_static_features(stations, static_features, st, en)
 
 
 START_YEAR = 2012
