### 2025-02-17 0af39d1
corrected path for camels_se
