### 2025-02-19 5c05de4
unzip .7z files of bull dataset
