### 2025-02-19 0b98b90
removed minor bug in usgs and correct cpu aquision
