### 2025-02-19 95eec74
updated paths for gsha
