+++ aqua_fetch/rr/_gsha.py
@@ -568,7 +568,7 @@ class GSHA(_RainfallRunoff):
             METEO_MAP[self.agency_of_stn(stn)],
             f'{stn}.csv') for stn in self.stations()]
 
-        cpus = self.processes or get_cpus() - 2
+        cpus = self.processes or max(get_cpus() - 2, 1)
         start = time.time()
 
         if self.verbosity:
@@ -667,7 +667,7 @@ class GSHA(_RainfallRunoff):
             "Storage",
             f'{stn}.csv') for stn in self.stations()]
 
-        cpus = self.processes or get_cpus() - 2
+        cpus = self.processes or max(get_cpus() - 2, 1)
         start = time.time()
 
         if self.verbosity: print(f"Reading storage vars for {len(self.stations())} stations using {cpus} cpus")
@@ -1240,7 +1240,7 @@ def streamflow_indices_all_stations(
         if verbosity: print(f"Reading from pre-existing {nc_path}")
         return xr.open_dataset(nc_path)
 
-    cpus = cpus or get_cpus() - 2
+    cpus = cpus or max(get_cpus() - 2, 1)
 
     start = time.time()
 
@@ -1278,12 +1278,12 @@ def streamflow_indices_stn(
 ) -> pd.DataFrame:
     fpath = os.path.join(
         ds_path,
-        "StreamflowIndices",
+        "StreamflowIndices_yearly",
         "StreamflowIndices",
         f'{stn}.csv')
 
     if not os.path.exists(fpath):
-        raise FileNotFoundError(f"{ds_path} for station {stn} not found")
+        raise FileNotFoundError(f"{fpath} not found")
 
     df = pd.read_csv(
         fpath, index_col=0,
@@ -1347,7 +1347,7 @@ def lc_vars_all_stns(
         if verbosity: print(f"Reading from pre-existing {nc_path}")
         return xr.open_dataset(nc_path)
 
-    cpus = cpus or get_cpus() - 2
+    cpus = cpus or max(get_cpus() - 2, 1)
 
     start = time.time()
 
@@ -1412,7 +1412,7 @@ def reservoir_vars_all_stns(
         if verbosity: print(f"Reading from pre-existing {nc_path}")
         return xr.open_dataset(nc_path)
 
-    cpus = cpus or get_cpus() - 2
+    cpus = cpus or max(get_cpus() - 2, 1)
 
     start = time.time()
 
@@ -1479,7 +1479,7 @@ def lai_all_stns(
         if verbosity: print(f"Reading from pre-existing {ds_path}")
         return pd.read_csv(os.path.join(ds_path, 'lai.csv'), index_col=0)
 
-    cpus = cpus or get_cpus() - 2
+    cpus = cpus or max(get_cpus() - 2, 1)
 
     start = time.time()
 
@@ -1517,7 +1517,7 @@ def lai_all_stns(
 
 # the dates for data to be downloaded 
 START_YEAR = 1979
-END_YEAR = 2023
+END_YEAR = 2024
 
 
 class Japan(_GSHA):
@@ -1675,32 +1675,66 @@ def download_daily_data(
 
     if verbosity:
         print(f"downloading daily data for {len(stations)} stations from {years[0]} to {years[-1]}")
-    
+
     if verbosity>1:
         print(f"Total function calls: {len(stations_)} with {cpus} cpus")
 
+    # It takes ~ 100 seconds to download data for a single station for all years (1979-2024) with 1 cpu
+
     start = time.time()
-    with cf.ProcessPoolExecutor(cpus) as executor:
-        results = executor.map(download_daily_stn_yr, stations_, years_)
+
+    if cpus == 1:
+        all_data = []
+        for idx, stn in enumerate(stations):
+            stn_data = []
[…]
