+++ tests/rr/test_hysets.py
@@ -37,29 +37,30 @@ gscad_path = '/mnt/datawaha/hyex/atr/gscad_database/raw'
 hy = HYSETS(path=gscad_path, verbosity=5)
 
 # because it takes very long time, we don't test with all the data
-test_dynamic_data(hy, 0.003, int(14425 * 0.003), 25202)
+test_dynamic_data(hy, 0.1, int(14425 * 0.1), 27028)
 
 test_static_data(hy, None, 14425)
 test_static_data(hy, 0.1, int(14425*0.1))
 
-test_all_data(hy, 3, 25202)
-test_all_data(hy, 3, 25202, True)
+test_all_data(hy, 2000, 27028)
+test_all_data(hy, 2000, 27028, True)
 
-test_attributes(hy, 28, 5, 14425)
+test_attributes(hy, 30, 20, 14425)
 
 test_fetch_dynamic_features(hy, random.choice(hy.stations()))
 test_fetch_dynamic_features(hy, random.choice(hy.stations()), True)
 
-test_fetch_dynamic_multiple_stations(hy, 3,  25202)
-test_fetch_dynamic_multiple_stations(hy, 3, 25202, True)
+test_fetch_dynamic_multiple_stations(hy, 3,  27028)
+test_fetch_dynamic_multiple_stations(hy, 3, 27028, True)
 
 test_fetch_static_feature(hy, random.choice(hy.stations()),
-                            14425, 28)
+                            14425, 30)
 
 test_st_en_with_static_and_dynamic(hy, random.choice(hy.stations()), yearly_steps=366)
 test_st_en_with_static_and_dynamic(hy, random.choice(hy.stations()), True, yearly_steps=366)
 
 test_selected_dynamic_features(hy)
+test_selected_dynamic_features(hy, True)
 
 test_coords(hy)
 
