+++ aqua_fetch/rr/_hysets.py
@@ -1,21 +1,45 @@
 
 import os
+from datetime import datetime
 from typing import Union, List, Dict, Tuple
 
 import numpy as np
 import pandas as pd
 
+try:
+    from netCDF4 import Dataset
+except (ModuleNotFoundError, ImportError):
+    pass
+
 from .utils import _RainfallRunoff
 from .._backend import shapefile, xarray as xr
-from ..utils import check_attributes
+from ..utils import check_attributes, download, _unzip
 
 from ._map import (
     observed_streamflow_cms,
+    observed_streamflow_mmd,
     min_air_temp,
     max_air_temp,
     mean_air_temp,
     total_precipitation,
     snow_water_equivalent,
+    mean_dewpoint_temperature_at_2m,
+    max_air_temp_with_specifier,
+    min_air_temp_with_specifier,
+    u_component_of_wind_at_10m,
+    v_component_of_wind_at_10m,
+    mean_daily_evaporation_with_specifier,
+    cloud_cover,
+    downward_longwave_radiation,
+    mean_thermal_radiation,
+    snow_density,
+    mean_daily_evaporation,
+    snowfall,
+    snowmelt,
+    mean_air_pressure,
+    solar_radiation,
+    net_longwave_radiation,
+    net_solar_radiation,
     )
 
 from ._map import (
@@ -29,7 +53,7 @@ from ._map import (
 class HYSETS(_RainfallRunoff):
     """
     database for hydrometeorological modeling of 14,425 North American watersheds
-    from 1950-2018 following the work of `Arsenault et al., 2020 <https://doi.org/10.1038/s41597-020-00583-2>`_
+    from 1950-2023 following the work of `Arsenault et al., 2020 <https://doi.org/10.1038/s41597-020-00583-2>`_
     The user must manually download the files, unpack them and provide
     the `path` where these files are saved.
 
@@ -106,20 +130,75 @@ class HYSETS(_RainfallRunoff):
 
     """
     doi = "https://doi.org/10.1038/s41597-020-00583-2"
-    url = "https://osf.io/rpc3w/"
-    Q_SRC = ['ERA5', 'ERA5Land', 'ERA5Land_SWE', 'Livneh', 'nonQC_stations', 'SCDNA', 'SNODAS_SWE']
-    SWE_SRC = ['ERA5Land_SWE', 'SNODAS_SWE']
-    OTHER_SRC = [src for src in Q_SRC if src not in ['ERA5Land_SWE', 'SNODAS_SWE']]
-    dynamic_features = [observed_streamflow_cms(), snow_water_equivalent(), 
-                        min_air_temp(), max_air_temp(), total_precipitation()]
+    url = {
+'HYSETS_watershed_boundaries.zip': 'https://osf.io/download/p8unw/',
+'HYSETS_watershed_properties.txt': 'https://osf.io/download/us795/',
+'HYSETS_2023_update_ERA5.nc': 'https://osf.io/download/fdnc8/',
+'HYSETS_2023_update_ERA5Land.nc': 'https://osf.io/download/4vt2s/',
+'HYSETS_2023_update_Livneh.nc': 'https://osf.io/download/4jgpt/',
+'HYSETS_2023_update_monthly_meteorological_data.nc': 'https://osf.io/download/sc4ge/',
+'HYSETS_2023_update_SNODAS.nc': 'https://osf.io/download/46wa7/',
+'HYSETS_2023_update_SCDNA.nc': 'https://osf.io/download/q8za6/',
+'HYSETS_2023_update_NRCAN.nc': 'https://osf.io/download/vfpre/',
[…]
