### 2025-02-24 b48c6eb
updated data paths for lamahce
