### 2025-02-24 26419f9
removed minor bug in estreams
