+++ aqua_fetch/rr/_hysets.py
@@ -204,41 +204,41 @@ class HYSETS(_RainfallRunoff):
         """
         parameters
         --------------
-            path : str
-                The path under which the data is to be saved or is saved already.
-                If the data is alredy downloaded then provide the path under which
-                HYSETS data is located. If None, then the data will be downloaded.
-                The data is downloaded once and therefore susbsequent
-                calls to this class will not download the data unless
-                ``overwrite`` is set to True.
-            sources : dict
-                sources for each dynamic feature. The keys should be dynamic features
-                and values should be sources. Available sources for the dynamic 
-                features are as below
-                    
-                    - 10m_u_component_of_wind: ['ERA5', 'ERA5Land']
-                    - 10m_v_component_of_wind: ['ERA5', 'ERA5Land']
-                    - 2m_dewpoint: ['ERA5', 'ERA5Land']
-                    - 2m_tasmax: ['NRCAN', 'Livneh', 'QC_stations', 'ERA5', 'nonQC_stations', 'ERA5Land', 'SCDNA']
-                    - 2m_tasmin: ['NRCAN', 'Livneh', 'QC_stations', 'ERA5', 'nonQC_stations', 'ERA5Land', 'SCDNA']
-                    - discharge: ['NRCAN', 'ERA5', 'ERA5Land', 'Livneh', 'nonQC_stations', 'SCDNA', 'SNODAS', 'QC_stations']
-                    - evaporation: ['ERA5', 'ERA5Land']
-                    - snow_density: ['ERA5', 'ERA5Land']
-                    - snow_evaporation: ['ERA5', 'ERA5Land']
-                    - snow_water_equivalent: ['ERA5', 'ERA5Land', 'SNODAS']
-                    - snowfall: ['ERA5', 'ERA5Land']
-                    - snowmelt: ['ERA5', 'ERA5Land']
-                    - surface_downwards_solar_radiation: ['ERA5', 'ERA5Land']
-                    - surface_downwards_thermal_radiation: ['ERA5', 'ERA5Land']
-                    - surface_net_solar_radiation: ['ERA5', 'ERA5Land']
-                    - surface_net_thermal_radiation: ['ERA5', 'ERA5Land']
-                    - surface_pressure: ['ERA5', 'ERA5Land']
-                    - surface_runoff: ['ERA5', 'ERA5Land']
-                    - total_cloud_cover: ['ERA5']
-                    - total_precipitation: ['NRCAN', 'Livneh', 'QC_stations', 'ERA5', 'nonQC_stations', 'ERA5Land', 'SCDNA']
+        path : str
+            The path under which the data is to be saved or is saved already.
+            If the data is alredy downloaded then provide the path under which
+            HYSETS data is located. If None, then the data will be downloaded.
+            The data is downloaded once and therefore susbsequent
+            calls to this class will not download the data unless
+            ``overwrite`` is set to True.
+        sources : dict
+            sources for each dynamic feature. The keys should be dynamic features
+            and values should be sources. Available sources for the dynamic 
+            features are as below
+                
+                - 10m_u_component_of_wind: ['ERA5', 'ERA5Land']
+                - 10m_v_component_of_wind: ['ERA5', 'ERA5Land']
+                - 2m_dewpoint: ['ERA5', 'ERA5Land']
+                - 2m_tasmax: ['NRCAN', 'Livneh', 'QC_stations', 'ERA5', 'nonQC_stations', 'ERA5Land', 'SCDNA']
[…]
