### 2025-02-25 838adc3
test selected dyn features for multiple stns
