### 2025-02-25 6581484
removed minor bug in estreams
