+++ tests/rr/test_gsha.py
@@ -279,7 +279,6 @@ test_dataset(ds,
               )
 
 
-
 ds = Spain(path=gscad_path, verbosity=3)
 
 test_dataset(ds,
