### 2025-02-26 2d24d59
refactored gsha class
