+++ aqua_fetch/rr/_gsha.py
@@ -434,11 +434,50 @@ class GSHA(_RainfallRunoff):
         """
         stations = self._get_stations(stations, agency)
 
-        lc_vars = lc_vars_all_stns(self.path)
+        lc_vars = self.lc_vars_all_stns()
         if isinstance(lc_vars, xr.Dataset):
             return lc_vars[stations]
         else:
             return {stn: lc_vars[stn] for stn in stations}
+
+    def lc_vars_all_stns(
+            self
+    ):
+        nc_path = os.path.join(self.path, 'lc_variables.nc')
+
+        if self.to_netcdf and os.path.exists(nc_path):
+            if self.verbosity: print(f"Reading from pre-existing {nc_path}")
+            return xr.open_dataset(nc_path)
+
+        cpus = self.processes or max(get_cpus() - 2, 1)
+
+        start = time.time()
+
+        stations = self.stations()
+        paths = [self.path for _ in range(len(stations))]
+
+        if self.verbosity: print(f"Reading landcover variables for {len(stations)} stations using {cpus} cpus")
+
+        with cf.ProcessPoolExecutor(cpus) as executor:
+
+            results = executor.map(
+                lc_variable_stn,
+                paths,
+                stations,
+            )
+
+        if self.verbosity: print(f"Time taken: {time.time() - start:.2f} seconds")
+
+        if self.to_netcdf:
+
+            encoding = {var: {'dtype': 'float32', 'zlib': True, 'complevel': 3} for var in stations()}
+
+            ds = xr.Dataset({stn: xr.DataArray(val) for stn, val in zip(stations, results)})
+            if self.verbosity: print(f"Saving to {nc_path}")
+            ds.to_netcdf(nc_path, encoding=encoding)
+        else:
+            ds = {stn: df for stn, df in zip(stations, results)}
+        return ds
 
     def reservoir_variables_stn(self, stn: str) -> pd.DataFrame:
         """
@@ -466,11 +505,50 @@ class GSHA(_RainfallRunoff):
         """
         stations = self._get_stations(stations, agency)
 
-        lc_vars = reservoir_vars_all_stns(self.path)
+        lc_vars = self.reservoir_vars_all_stns()
         if isinstance(lc_vars, xr.Dataset):
             return lc_vars[stations]
         else:
             return {stn: lc_vars[stn] for stn in stations}
+
+    def reservoir_vars_all_stns(
+            self
+    ):
+        nc_path = os.path.join(self.path, 'reservoir_variables.nc')
+
+        if self.to_netcdf and os.path.exists(nc_path):
+            if self.verbosity: print(f"Reading from pre-existing {nc_path}")
+            return xr.open_dataset(nc_path)
+
+        cpus = self.processes or max(get_cpus() - 2, 1)
+
+        start = time.time()
+
+        stations = self.stations()
+        paths = [self.path for _ in range(len(stations))]
+
+        if self.verbosity: print(f"Reading reservoir variables for {len(stations)} stations using {cpus} cpus")
+
+        with cf.ProcessPoolExecutor(cpus) as executor:
+
+            results = executor.map(
+                reservoir_vars_stn,
+                paths,
+                stations,
+            )
+
+        if self.verbosity: print(f"Time taken: {time.time() - start:.2f} seconds")
+
+        if self.to_netcdf:
+
+            encoding = {var: {'dtype': 'float32', 'zlib': True, 'complevel': 3} for var in stations}
+
+            ds = xr.Dataset({stn: xr.DataArray(val) for stn, val in zip(stations, results)})
[…]
