### 2025-02-26 4890fee
remove duplicated static features from cabra
