### 2025-02-26 05cac9c
calculate mean air temp for cabra
