+++ tests/rr/test_brazil.py
@@ -23,7 +23,7 @@ from utils import (
 
 for source in ['era5', 'ref', 'ens']:
     dataset = CABra(path=gscad_path, met_src=source, verbosity=4)
-    test_dataset(dataset, 735, 10957, 87, 12)
+    test_dataset(dataset, 735, 10957, 87, 13)
 
 
 ds_br = CAMELS_BR(path=os.path.join(gscad_path, 'CAMELS'), verbosity=3)
