+++ readme.md
@@ -163,7 +163,7 @@ mg_data_ohe.shape
 | WaterBenchIowa | 125                    |                         | 3                        | 7                       | 2011 - 2018       | Iowa (USA)                                  | [Demir et al., 2022](https://doi.org/10.5194/essd-14-5605-2022)                                             |
 | CCAM           | 102                    |                         | 16                       | 124                     | 1990 - 2020       | China                                       | [Hao et al., 2021](https://doi.org/10.5194/essd-13-5591-2021)                                               |
 | RRLuleaSweden  | 1                      |                         | 2                        | 0                       | 2016 - 2019       | Lulea (Sweden)                              | [Broekhuizen et al., 2020](https://doi.org/10.5194/hess-24-869-2020)                                        |
-| CABra          | 735                    |                         | 12                       | 97                      | 1980 - 2010       | Brazil                                      | [Almagro et al., 2021](https://doi.org/10.5194/hess-25-3105-2021)                                           |
+| CABra          | 735                    |                         | 13                       | 87                      | 1980 - 2010       | Brazil                                      | [Almagro et al., 2021](https://doi.org/10.5194/hess-25-3105-2021)                                           |
 | HYPE           | 561                    |                         | 9                        | 3                       | 1985 - 2019       | Costa Rica                                  | [Arciniega-Esparza and Birkel, 2020](https://zenodo.org/records/4029572)                                    |
 | Ireland        | 464                    |                         | 27                       | 35                      | 1992 - 2020       | Ireland                                     | [EPA Ireland](https://epawebapp.epa.ie)                                                                     |
 | Spain          | 889                    |                         | 27                       | 35                      | 1979 - 2020       | Spain                                       | [ceh-flumen64](https://ceh-flumen64.cedex.es)                                                               |
