### 2025-02-26 26e227e
calculate more features for camels_aus
