### 2025-03-02 aedc225
removed bug in downloading polish data
