+++ docs/source/rr.rst
@@ -3,184 +3,371 @@ Rainfall Runoff datasets
 This section include datasets which can be used for rainfall runoff modeling.
 They all contain observed streamflow and meteological data as time series.
 These are named as ``dynamic features``. The physical catchment properties
-are included as ``static features``. Although each data source has a dedicated
-class, however :py:class:`aqua_fetch.rr.RainfallRunoff` class can be used to access all the datasets.
+are included as ``static features`` as tabular data, where each row corresponds 
+to one catchment and each column to one static feature.
+
+In addition to published datasets, this package introduces 10 new datasets for 
+rainfall-runoff modeling. These datasets have not yet been published but follow 
+the CAMELS dataset series convention. They include Ireland, Finland, Italy, Poland, 
+Portugal, Japan, Thailand, Arcticnet, Spain, and the USGS. The observed streamflow 
+data are sourced from the national meteorological or hydrological websites of the 
+respective countries. Catchment boundaries and meteorological data for Ireland, 
+Finland, Italy, Poland, and Portugal are obtained from EStreams (`Nascimento et al., 2024 <https://doi.org/10.1038/s41597-024-03706-1>`_), 
+and similarly for Japan, Thailand, Arcticnet, and Spain from GSHA (`Peirong et al., 2023 <https://doi.org/10.5194/essd-16-1559-2024>`_). 
+For USGS, the catchment boundaries are sourced from HYSETS (`Arsenault et al., 2020 <https://doi.org/10.1038/s41597-020-00583-2>`_).
+
+Although each data source has a dedicated, however
+all datasets listed in Table \ref{tbl:table1} are accessible via the :py:class:`aqua_fetch.rr.RainfallRunoff` 
+class, which allows for a unified and consistent approach to each dataset. The class 
+provides several methods to access static features, dynamic features, or catchment 
+boundaries. Although the raw data files for each dataset may come in different formats, 
+the methods to access these features through the :py:class:`aqua_fetch.rr.RainfallRunoff` class remain the same. 
+Individual classes for each dataset are also available and may offer more control to 
+users over specific datasets. However, for most cases, the use of the :py:class:`aqua_fetch.rr.RainfallRunoff` 
+class will suffice.
+
+The naming and units of dynamic features in each dataset may vary. However, we have 
+standardized these features using the formula `name_unit_specifier` for each dynamic 
+feature across all datasets. In this formula, the specifier can indicate the source 
+(such as ERA5 or MSWEP for precipitation), the method used to calculate the feature 
+(like makkink or penman for evapotranspiration), or the aggregation type (min, max, mean). 
+For example, a precipitation dynamic feature from MSWEP would be labeled as pcp_mm_mswep. 
+This approach ensures that feature names are representative and understandable. 
+Dynamic features for which this method is inapplicable retain their original names.
+
[…]
