### 2025-03-19 28ce8fd
updated table in readme
