+++ aqua_fetch/rr/_camels.py
@@ -96,7 +96,7 @@ class CAMELS_US(_RainfallRunoff):
     >>> dataset.dynamic_features
     # get only selected dynamic features
     >>> _, df = dataset.fetch(1, as_dataframe=True,
-    ... dynamic_features=['prcp(mm/day)', 'srad(W/m2)', 'tmax(C)', 'tmin(C)', 'Flow']).unstack()
+    ... dynamic_features=['pcp_mm', 'solrad_wm2', 'airtemp_C_max', 'airtemp_C_min', 'q_cms_obs']).unstack()
     >>> df.shape
     (12784, 5)
     # get names of available static features
@@ -112,8 +112,21 @@ class CAMELS_US(_RainfallRunoff):
 
     """
     DATASETS = ['CAMELS_US']
-    url = "https://ral.ucar.edu/sites/default/files/public/product-tool/camels-catchment-attributes-and-meteorology-for-large-sample-studies-dataset-downloads/basin_timeseries_v1p2_metForcing_obsFlow.zip"
-    catchment_attr_url = "https://ral.ucar.edu/sites/default/files/public/product-tool/camels-catchment-attributes-and-meteorology-for-large-sample-studies-dataset-downloads/camels_attributes_v2.0.zip"
+
+    url = {
+        'camels_attributes_v2.0.pdf': 'https://gdex.ucar.edu/dataset/camels/file/',
+        'camels_attributes_v2.0.xlsx': 'https://gdex.ucar.edu/dataset/camels/file/',
+        'camels_clim.txt': 'https://gdex.ucar.edu/dataset/camels/file/',
+        'camels_geol.txt': 'https://gdex.ucar.edu/dataset/camels/file/',
+        'camels_hydro.txt': 'https://gdex.ucar.edu/dataset/camels/file/',
+        'camels_name.txt': 'https://gdex.ucar.edu/dataset/camels/file/',
+        'camels_soil.txt': 'https://gdex.ucar.edu/dataset/camels/file/',
+        'camels_topo.txt': 'https://gdex.ucar.edu/dataset/camels/file/',
+        'camels_vege.txt': 'https://gdex.ucar.edu/dataset/camels/file/',
+        'readme.txt': 'https://gdex.ucar.edu/dataset/camels/file/',
+        'basin_timeseries_v1p2_metForcing_obsFlow.zip': 'https://gdex.ucar.edu/dataset/camels/file/',
+        'basin_set_full_res.zip': 'https://gdex.ucar.edu/dataset/camels/file/',
+    }
 
     folders = {'basin_mean_daymet': f'basin_mean_forcing{SEP}daymet',
                'basin_mean_maurer': f'basin_mean_forcing{SEP}maurer',
@@ -157,20 +170,37 @@ class CAMELS_US(_RainfallRunoff):
 
         super().__init__(path=path, name="CAMELS_US", **kwargs)
 
-        self.path = path
+        if not os.path.exists(self.path):
+            os.makedirs(self.path)
 
-        if os.path.exists(self.path):
-            if self.verbosity: print(f"dataset is already downloaded at {self.path}")
-        else:
-            download(self.url, os.path.join(self.camels_dir, f'CAMELS_US{SEP}CAMELS_US.zip'))
-            download(self.catchment_attr_url, os.path.join(self.camels_dir, f"CAMELS_US{SEP}catchment_attrs.zip"))
-            unzip(self.path)
+        for fname, url in self.url.items():
 
-        self.attr_dir = os.path.join(self.path, f'catchment_attrs{SEP}camels_attributes_v2.0')
-        self.dataset_dir = os.path.join(self.path, f'CAMELS_US{SEP}basin_dataset_public_v1p2')
+            fpath = os.path.join(self.path, fname)
[…]
