### 2025-04-07 #7705 @ethanwhite
👋 @jmp75 @changliao1025 - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
