### 2025-04-22 hyex-research/AquaFetch#7 @jeiloh
JOSS review
Hi, this issue has been created to provide comments as part of my review of the AquaFetch software package for the Journal of Open Source Software (openjournals/joss-reviews#8051).

The AquaFetch package is a effort aimed at simplifying and streamlining access, harmonization, and preprocessing of diverse water resource datasets. It helps to reduce the burden of manual dataset management by providing a unified Python interface for rainfall-runoff, water-quality, and wastewater-treatment data from publicly and freely available sources.

I will be adding comments, questions, and general suggestions to this thread as I proceed with my review. Feel free to respond or address them.

## April 22
### Clarify unique value
- The manuscript presents AquaFetch as a valuable contribution to the water resource dataset management ecosystem. While the extensive dataset coverage (70 datasets) is impressive, I would appreciate if the authors could expand on what makes AquaFetch uniquely valuable beyond this coverage. I am particularly wondering about the comparison with neuralhydrology (https://neuralhydrology.readthedocs.io/en/latest/api/neuralhydrology.datasetzoo.html), which shows data processing for several datasets including CAMELS and Caravan. How does AquaFetch differ from those functions?

### Enhance community contribution guidance
- It may be helpful to include a brief example or a link to a CONTRIBUTING guide that illustrates the typical steps for adding a new dataset (e.g., directory layout, metadata file, code stub, unit test, and PR submission).

### Corrections
- **Review Installation instructions**
  I found the example uses `AquaFetch[all]` but the package metadata is `aqua‑fetch`. It should be:
    ```bash
    pip install "aqua‑fetch[all] @ git+https://github.com/hyex-research/AquaFetch.git"
    ```

- **Correct `fetch(...).unstack()` usage [example](https://aquafetch.readthedocs.io/en/latest/rr.html#aqua_fetch.rr.RainfallRunoff)**
  Calling `.unstack()` directly on `dataset.fetch(..., as_dataframe=True)` raises an `AttributeError` because `fetch` returns a tuple (e.g., `(meta, df)`).

### Paper
- **Correct typographical errors**
  - **Location:** Section header on p. 2 (“Implementation and Archiecture”)
  - Fix “Archiecture” → **Architecture**.

- **Fix incorrect import name in example**
  - **Location:** Code snippet (p. 3, lines 74–75)
  - Change `RainfallRUnoff` → `RainfallRunoff`.
  - Change `SWatch` → `SWatCh`.
