+++ paper/paper.md
@@ -50,8 +50,8 @@ Ultimately, we hope this package will foster the development of benchmark datase
 The AquaFetch package is structured using both functional and Object-Oriented Programming (OOP) designs. The OOP design is employed for handling more complex datasets within the `rr` and `wq` submodules, while simpler datasets are managed through a functional interface. All datasets in `rr` submodule can be accessed via `RainfallRunoff` class thus offering a unified interface. The package's code is logically organized, with `rr`, `wq`, and `wwt` subdirectories present in both the source code (aqua_fetch) and tests directories. All public classes and functions are accessible from the parent directory, allowing for straightforward imports as shown below:
 
 ```python
-    from aqua_fetch import RainfallRUnoff
-    from aqua_fetch import SWatch
+    from aqua_fetch import RainfallRunoff
+    from aqua_fetch import SWatCh
     from aqua_fetch import mg_degradation
 ```
 
