+++ readme.md
@@ -36,7 +36,7 @@ The above code will install minimal depencies required to use the library which 
 numpy, pandas and requests. To install the library with full list of dependencies use the
 ``all`` option during installation.
 
-	python -m pip install "AquaFetch[all] @ git+https://github.com/hyex-research/AquaFetch.git"
+	python -m pip install "aqua-fetch[all] @ git+https://github.com/hyex-research/AquaFetch.git"
 
 This will install addtional optional depencdies which include xarray, pyshp, netCDF and easy_mpl.
 
@@ -185,16 +185,16 @@ mg_data_ohe.shape
 |---------------------------|-------------------|-------------------|---------------------------|------------------------------------------------------------------------------|
 | SWatCh                    | 24                | 1960 - 2022       | Global                    | [Lobke et al., 2022](https://doi.org/10.5194/essd-14-4667-2022)              |
 | GRQA                      | 42                | 1898 - 2020       | Global                    | [Virro et al., 2021](https://essd.copernicus.org/articles/13/5483/2021/)     |
-| Quadica                   |                   | 1950 - 2018       | Germany                   | [Ebeling et al., 2022 ](https://essd.copernicus.org/articles/14/3715/2022/)  |
+| Quadica                   | 10                | 1950 - 2018       | Germany                   | [Ebeling et al., 2022 ](https://essd.copernicus.org/articles/14/3715/2022/)  |
 | RC4USCoast                | 21                | 1850 - 2020       | USA                       | [Gomez et al., 2022](https://essd.copernicus.org/articles/15/2223/2023/)     |
-| Busan Beach               |                   | 2018 - 2019       | Busan, South Korea        | [Jang et al., 2021](https://doi.org/10.1016/j.watres.2021.117001)                   |
+| Busan Beach               | 14                | 2018 - 2019       | Busan, South Korea        | [Jang et al., 2021](https://doi.org/10.1016/j.watres.2021.117001)                   |
 | Ecoli Mekong River        | 10                | 2011 - 2021       | Mekong river (Houay Pano) | [Boithias et al., 2022](https://essd.copernicus.org/articles/14/2883/2022/)  |
 | Ecoli Mekong River (Laos) | 10                | 2011 - 2021       | Mekong River (Laos)       | [Boithias et al., 2022](https://essd.copernicus.org/articles/14/2883/2022/)  |
 | Ecoli Houay Pano (Laos)   | 10                | 2011 - 2021       | Houay Pano (Laos)         | [Boithias et al., 2022](https://essd.copernicus.org/articles/14/2883/2022/)  |
-| CamelsChem                | 18                | 1980 - 2018       | Conterminous USA          | [Sterle et al., 2024](https://doi.org/10.5194/hess-28-611-2024)              |
-| GRiMeDB                   | 18                | -                 | Global                    | [Stanley et al., 2023](https://doi.org/10.5194/essd-15-2879-2023)            |
+| CamelsChem                | 28                | 1980 - 2018       | Conterminous USA          | [Sterle et al., 2024](https://doi.org/10.5194/hess-28-611-2024)              |
+| GRiMeDB                   | 1                 | 1973 - 2021       | Global                    | [Stanley et al., 2023](https://doi.org/10.5194/essd-15-2879-2023)            |
 | SanFrancisco Bay          | 18                | 1969 - 2015       | Sans Francisco Bay (USA)  | [Cloern et al., 2017](https://doi.org/10.1002/lno.10537)                     |
[…]
