### 2025-04-27 bf16481
updated example for CAMELS_AUS for v2
