### 2025-04-28 #8051 @ethanwhite
👋 @jeiloh & @tamnva! Thanks again for agreeing to review this paper. I can see that both of you have created your checklists and that @jeiloh is already making progress. Much appreciated! Since we're now two weeks into the review process, I just wanted to check in and see if either of you have any questions or have run into any issues that I can help out with as you work on your reviews. If so, let me know!
