### 2025-04-29 hyex-research/AquaFetch#8 @tamnva
JOSS comment (from tamnva)
Dear authors of AquaFetch,

First of all, I would like to thank the authors for creating this package for the community. This issue was created to store my comments for the manuscript submitted to JOSS (https://github.com/openjournals/joss-reviews/issues/8051). Please find below the reasons for the unchecked items in JOSS reviewe.

## General checks

1. Contribution and authorship: please explain the contribution to the package developemnt of all coauthors and the list of people that committed to the package github repo

## Functionality
1. Installation/Installation instructions: The dev_requirements.txt and requirements.txt file should list of dependency packages, including the python version.

2. Functionality
- The code provided in the Usage section of the README file is quite confusing and does not work, mixing of both python script and outputs.
- The package name is "aqua-fetch" but when import in python using the name "aqua_fetch" (for example from aqua_fetch import RainfallRunoff). Please check the name following standard naming convention, for example using PEP 8 https://peps.python.org/pep-0008/

Sorry I cannot further review the environment as the I have issue with importing the software, could the authors please have a look at this so I could continue my review

Note: attached is the environment that I am working with and the error when I importing the function

>>> from aqua_fetch import RainfallRunoff
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "C:\Users\user_name\AppData\Local\anaconda3\envs\aqua1\Lib\site-packages\aqua_fetch\__init__.py", line 7, in <module>
    from .rr import _RainfallRunoff
  File "C:\Users\user_name\AppData\Local\anaconda3\envs\aqua1\Lib\site-packages\aqua_fetch\rr\__init__.py", line 14, in <module>
    from .utils import _RainfallRunoff
  File "C:\Users\user_name\AppData\Local\anaconda3\envs\aqua1\Lib\site-packages\aqua_fetch\rr\utils.py", line 31, in <module>
    class _RainfallRunoff(Datasets):
  File "C:\Users\user_name\AppData\Local\anaconda3\envs\aqua1\Lib\site-packages\aqua_fetch\rr\utils.py", line 881, in _RainfallRunoff
    ) -> plt.Axes:
         ^^^^^^^^
AttributeError: 'NoneType' object has no attribute 'Axes'
