### 2025-05-02 hyex-research/AquaFetch#8 @jeiloh
The command `pip install aqua-fetch` from PyPI fails. Initially, it throws AttributeError: 'NoneType' object has no attribute 'Axes'. I think this is because `matplotlib` is not installed as a base dependency.

After manually installing `matplotlib`, a different error appears on import: AttributeError: 'NoneType' object has no attribute 'Dataset'. This seems `xarray` is also missing from the base dependencies required for import.
Confirmed that installing via `pip install "aqua-fetch[all] @ git+https://github.com/hyex-research/AquaFetch.git"` worked correctly.

@AtrCheema Could the PyPI package be updated, or the base dependency specification (install_requires) be reviewed to ensure pip install aqua-fetch provides a working installation out-of-the-box?
