### 2025-05-12 #8051 @ethanwhite
@tamnva - thanks! Sorry I missed that highly important detail when skimming through the issue this morning.

@jeiloh - I also saw that I'd missed your review issue when glancing through. Are you currently waiting on anything or do you consider your review to be complete pending changes from the author?

 @AtrCheema - the PyPI version of the package is currently not installing cleanly for either of the reviewers. Can you make a new release from your current main PyPI branch (which does appear to be installing cleanly) so that reviewers can test through that installation approach?
