### 2025-05-12 9725f9a
make matplotlib optional dependency
