+++ aqua_fetch/rr/__init__.py
@@ -771,7 +771,7 @@ class RainfallRunoff(object):
             ax: plt_Axes = None,
             show: bool = True,
             **kwargs
-    ) -> plt.Axes:
+    ):
         """
         plots catchment boundaries
 
