### 2025-05-12 6c141fe
dont throw error if xarray is not installed
