### 2025-05-12 #8051 @AtrCheema
Hi all, thanks for the highlight. The error occurs because matplotlib is not installed. We have resolved this issue in the most recent commits; however, these changes have not yet been pushed to PyPI. This is because we plan to make a new release with significant improvements and changes, and we want to avoid cluttering the PyPI repository with frequent minor updates.

Would it be possible for you to install the package using `pip install aqua_fetch['all']`? This will ensure that both matplotlib and xarray are installed.
