### 2025-05-12 #8051 @ethanwhite
> Would it be possible for you to install the package using pip install aqua_fetch['all']? This will ensure that both matplotlib and xarray are installed.

This seems like a reasonable solution to me. @tamnva - does that work for you?

(NB: if using zsh the command needs to be `pip install 'aqua_fetch['all']'` due to how zsh handles square brackets
