### 2025-05-12 #8051 @jeiloh
Hi @ethanwhite and @AtrCheema,

I have successfully installed the package using `pip install aqua_fetch['all'] `and have proceeded with the review (https://github.com/hyex-research/AquaFetch/issues/7). I am currently awaiting the author's response, particularly concerning two main aspects:

1. Clarification of the software's unique value.
2. Enhancement of the guidance for community contributions.

I plan to complete testing and address any other outstanding points within the next two weeks.
