### 2025-05-18 hyex-research/AquaFetch#8 @tamnva
Dear authors of AquaFetch,

Please find below my comments (in addition to the comments in my first post)

###General checks
- Substantial scholarly effort: Could the authors provide some examples where the software has already been cited in academic papers?
- Whether the software is sufficiently useful that it is likely to be cited by other researchers working in this domain: I think this is usefull in the case that someone wants to automatically download the data and convert to specific format for further analysis (e.g., large sample analysis or machine learning modelling exercise).

### Functionality
- There are errors in importing functions, running the examples codes (either in the readme.md file or in the documentation, please find the error message in the attached files, the environment file was also attached)

### Documentation
- Please update the documentation and rerun the python scripts included in this documentation to make sure that they works, there are also some error in the documentations (https://aquafetch.readthedocs.io/en/latest/auto_examples/index.html)
-  Automated tests: I think the current testing using the github workflow (https://github.com/hyex-research/AquaFetch/blob/master/.github/workflows/draft-pdf.yml) is for testing the creation of the pdf file for JOSS, not testing the package functionality itself, is there a way to include testing of the package functionality within this workflow?

[error_1.txt](https://github.com/user-attachments/files/20271471/error_1.txt)
[error_2.txt](https://github.com/user-attachments/files/20271473/error_2.txt)
[error_3.txt](https://github.com/user-attachments/files/20271472/error_3.txt)
[error_4.txt](https://github.com/user-attachments/files/20271476/error_4.txt)
[error_5.txt](https://github.com/user-attachments/files/20271475/error_5.txt)
[error_6.txt](https://github.com/user-attachments/files/20271474/error_6.txt)
[error_7.txt](https://github.com/user-attachments/files/20271470/error_7.txt)
