### 2025-05-22 hyex-research/AquaFetch#7 @jeiloh
## May 21
### API Documentation Examples Need Updating
Some example code snippets in the API documentation appear to use outdated function calls or import paths that don't align with the current library version. For example,

An import like `from water.datasets import Swatch` might be outdated.
The example `from aqua_fetch import mg_degradation; mg_data, catalyst_encoder, anion_encoder = mg_degradation()` causes a `ValueError: not enough values to unpack (expected 3, got 2) `.

Please review and update the API documentation examples to ensure they reflect current usage.

### SWatCh Dataset Download Failure
An `UnboundLocalError` occurs when attempting to download the SWatCh dataset using SWatCh(). This seems to be related to the handling of a specific file entry from the Zenodo record.

> downloading https://zenodo.org/record/6484939 to ./my_swatch_data/SWatCh
> Total data to be downloaded is 1.52547 GB
> Title: The Surface Water Chemistry (SWatCh) database
> Keywords: water chemistry, water quality, water, acidification, acid rain, dataset, database, data, global, global coverage, global extent
> Publication date: 2021-02-24
> DOI: 10.5281/zenodo.6484939
> Total size: 1454.8 MB
>
> Link: https://zenodo.org/api/records/6484939/files/SWatCh_Sites_v2.sbx/content   size: 0.0 MB
> 0% of 0.0 MB downloaded
> 100% of 0.0 MB downloaded
>   Download error.
>   Too many errors.

The `UnboundLocalError` for the filename variable seems to occur because the download logic encounters an issue with the `SWatCh_Sites_v2.sbx` file
