### 2025-05-22 #8051 @jeiloh
Hi @AtrCheema and @ethanwhite,

Please find my review of the software's functionality at https://github.com/hyex-research/AquaFetch/issues/7

Overall, the software provides useful tools for users to download water-related datasets. However, as mentioned earlier, there are existing tools that offer similar functionality, particularly for rainfall-runoff datasets. Clearer differentiation of aqua-fetch from these existing solutions would strengthen the software's positioning.
The paper mentions that aqua-fetch offers various datasets including water quality and wastewater treatment plant data. While I appreciate the breadth of coverage, I would like to see more discussion of the synergies between these diverse datasets, as they serve quite different research purposes and user communities.
I still believe aqua-fetch has strong potential as a centralized venue for hosting and providing water-related datasets. To realize this potential, community participation and engagement will be crucial. The software would benefit from enhanced community contribution guidance. For instance, clearer documentation on how researchers can contribute their datasets—including submission procedures, data formatting requirements, and update processes—would encourage broader adoption and participation.
