### 2025-06-14 cce914f
removed bug in fetching data without xr
