### 2025-06-15 24331c0
updated tests for minimal dependencies
