+++ tests/rr/utils.py
@@ -7,7 +7,7 @@ import numpy as np
 import pandas as pd
 from aqua_fetch._backend import xarray as xr
 
-from aqua_fetch._backend import plt
+from aqua_fetch._backend import plt, netCDF4, shapefile
 
 logger = logging.getLogger(__name__)
 
@@ -181,7 +181,7 @@ def test_selected_dynamic_features(dataset, as_dataframe=False):
 
     if as_dataframe:
         data = data.unstack()
-        assert data.shape[1] == 2
+        assert data.shape[1] == 2, f"data.shape is {data.shape}, expected 2 dynamic features"
     else:
         assert len(data.dynamic_features) == 2, len(data.dynamic_features)
 
@@ -267,10 +267,11 @@ def check_dataframe(
         #         _stn_data_len = len(stn_data.iloc[stn_data.index.get_level_values('dynamic_features') == dyn_attr])
         #         assert _stn_data_len>=stn_data_len, f"{col} for {dataset.name} is not of length {stn_data_len}"
         stn_data = df[col].unstack()
+        desired_shape = (data_len, len(dataset.dynamic_features))
         # data for each station must minimum be of this shape
-        msg = f"""for {col} station of {dataset.name} the shape is {stn_data.shape}"""
+        msg = f"""for {col} station of {dataset.name} the shape is {stn_data.shape} and not {desired_shape}"""
         if raise_len_error:
-            assert stn_data.shape == (data_len, len(dataset.dynamic_features)), msg
+            assert stn_data.shape == desired_shape, msg
         else:
             logger.warning(msg)
 
@@ -363,24 +364,28 @@ def test_dataset(dataset, num_stations, dyn_data_len, num_static_attrs, num_dyn_
                  st="20040101", en="20041231",
                  has_q: bool = True,
                  ):
-    # check that dynamic attribues from all data can be retrieved.
-    test_dynamic_data(dataset, 'all', num_stations, dyn_data_len)
-    if test_df:
-        test_dynamic_data(dataset, 'all', num_stations, dyn_data_len, as_dataframe=True)
+    
+    # if netCDF4 is not None:
+    #     # check that dynamic attribues from all data can be retrieved.
+    #     test_dynamic_data(dataset, 'all', num_stations, dyn_data_len)
+    # if test_df:
+    #     test_dynamic_data(dataset, 'all', num_stations, dyn_data_len, as_dataframe=True)
 
-    # check that dynamic data of 10% of stations can be retrieved
-    test_dynamic_data(dataset, 0.1, int(num_stations * 0.1), dyn_data_len,
-                      raise_len_error=raise_len_error)
-    if test_df:
-        test_dynamic_data(dataset, 0.1, int(num_stations * 0.1), dyn_data_len, True,
-                          raise_len_error=raise_len_error)
+    # if netCDF4 is not None:
+    #     # check that dynamic data of 10% of stations can be retrieved
+    #     test_dynamic_data(dataset, 0.1, int(num_stations * 0.1), dyn_data_len,
+    #                   raise_len_error=raise_len_error)
+    # if test_df:
+    #     test_dynamic_data(dataset, 0.1, int(num_stations * 0.1), dyn_data_len, True,
+    #                       raise_len_error=raise_len_error)
 
-    test_static_data(dataset, 'all', num_stations)  # check that static data of all stations can be retrieved
+    # test_static_data(dataset, 'all', num_stations)  # check that static data of all stations can be retrieved
 
-    test_static_data(dataset, 0.1,
-                     int(num_stations * 0.1))  # check that static data of 10% of stations can be retrieved
+    # test_static_data(dataset, 0.1,
[…]
