### 2025-06-15 8b2398e
allow RainfallRunoff to access attributes from datasets
