### 2025-06-15 9d07b1c
updated docs and use st en while reading dynamic data
