+++ docs/source/conf.py
@@ -14,7 +14,7 @@ project = 'AquaFetch'
 copyright = '2025, Ather Abbas'
 author = 'Ather Abbas'
 
-release = "1.0.0rc2"
+release = "1.0.0rc3"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
