### 2025-06-15 fb0c674
updated docs and tests for gsha and npctr
