### 2025-06-15 3ebf7fd
removed bug in gw_punjab
