+++ CONTRIBUTING.rst
@@ -21,10 +21,23 @@ adding new dataset
 If you find a new dataset which is not covered in this dataset, you can add it by following
 these steps
 
-- For rainfall-runoff datasets, the new dataset has both observed daily (or sub-daily) streamflow and catchment boundaries available.
-- Now consider the name of new dataset as ``newdatasoruce`` and complete the remaining steps
-- create a file ``_newdatasource.py`` in ``aqua_fetch/submodule`` folder
-- write class ``NewDataSource`` in this file which must inherit from ``Datasets`` or ``_RainfallRunoff`` (for rainfall-runoff case) class
+rainfall-runoff
+==================
+Consider that you want to add a new rainfall-runoff dataset named ``NewDataSource``.
+- create a new file in ``aqua_fetch/rr/_newdatasource.py`` file
+- create a new class in this file which must inherit from ``_RainfallRunoff`` class
+- implement these methods ``__init__``, method which must download the data. The signature of these methods/attributes must be same as the parent class.
+    - ``__init__``  -> within this method, the logic to download the data must be implemented. The data must be downloaded within `path`
+    - ``_read_dynamic`` -> must return a dictionary with keys as station ids and values as pandas DataFrame with columns as names of dynamic_features and datetime index
+    - ``_read_static``  -> must return a pandas DataFrame with columns as names of static_features and index as station ids
+    - ``end``  -> must return the end date of the dataset
+    - ``start`` -> must return the start date of the dataset
+    - ``dynamic_features`` -> must return a list of dynamic features that are available in the dataset
+    - ``static_features`` -> must return a list of static features that are available in the dataset
+    - ``stations`` -> must return a list of station ids that are available in the dataset
+- The implementation of following methods is optional but recommended
+    - ``stn_coords``
+    - ``q_mmd``
 - implment ``__init__``, ``stn_coords``, and ``stations`` methods in this class. Check for documentation of these methods in ``Datasets`` class
 - import this class in ``__init__.py`` file in aqua_fetch folder and change the ``ALL_DATASETS`` lists to include this new class
 - Write tests for this new data source in tests directory where you compare the files generated by this new data source with the raw files.
