### 2025-06-25 1e6270c
added CAMELS_NZ dataset
