+++ aqua_fetch/rr/_camels.py
@@ -1,6 +1,7 @@
 import os
 import json
 import glob
+import time
 import shutil
 import warnings
 import concurrent.futures as cf
@@ -51,7 +52,8 @@ from ._map import (
     catchment_area,
     gauge_latitude,
     gauge_longitude,
-    slope
+    slope,
+    gauge_elevation_meters,
     )
 
 # directory separator
@@ -89,15 +91,15 @@ class CAMELS_US(_RainfallRunoff):
     >>> data.index.names == ['time', 'dynamic_features']
      True
     # get data by station id
-    >>> _, df = dataset.fetch(stations='11478500', as_dataframe=True).unstack()
-    >>> df.shape
+    >>> _, df = dataset.fetch(stations='11478500', as_dataframe=True)
+    >>> df.unstack().shape
     (12784, 8)
     # get names of available dynamic features
     >>> dataset.dynamic_features
     # get only selected dynamic features
     >>> _, df = dataset.fetch(1, as_dataframe=True,
-    ... dynamic_features=['pcp_mm', 'solrad_wm2', 'airtemp_C_max', 'airtemp_C_min', 'q_cms_obs']).unstack()
-    >>> df.shape
+    ... dynamic_features=['pcp_mm', 'solrad_wm2', 'airtemp_C_max', 'airtemp_C_min', 'q_cms_obs'])
+    >>> df.unstack().shape
     (12784, 5)
     # get names of available static features
     >>> dataset.static_features
@@ -107,8 +109,8 @@ class CAMELS_US(_RainfallRunoff):
     (102272, 10)  # remember this is multi-indexed DataFrame
     # If we want to get both static and dynamic data 
     >>> static, dynamic = dataset.fetch(stations='11478500', static_features="all", as_dataframe=True)
-    >>> static.shape, dynamic.shape
-    ((1, 59), (102272, 1))
+    >>> static.shape, dynamic.unstack().shape
+    ((1, 59), (12784, 8))
 
     """
     DATASETS = ['CAMELS_US']
@@ -141,8 +143,8 @@ class CAMELS_US(_RainfallRunoff):
 
     def __init__(
             self,
+            path:Union[str, os.PathLike]=None,
             data_source: str = 'basin_mean_daymet',
-            path=None,
             **kwargs
     ):
 
@@ -382,15 +384,15 @@ class CAMELS_GB(_RainfallRunoff):
     >>> len(stns)
     671
     # get data by station id
-    >>> _, df = dataset.fetch(stations='97002', as_dataframe=True).unstack()
-    >>> df.shape
+    >>> _, df = dataset.fetch(stations='97002', as_dataframe=True)
+    >>> df.unstack().shape
     (16436, 10)
     # get names of available dynamic features
     >>> dataset.dynamic_features
     # get only selected dynamic features
     >>> _, df = dataset.fetch(1, as_dataframe=True,
-    ... dynamic_features=['windspeed_mps', 'airtemp_C_mean', 'pet_mm', 'pcp_mm', 'q_cms_obs']).unstack()
-    >>> df.shape
+    ... dynamic_features=['windspeed_mps', 'airtemp_C_mean', 'pet_mm', 'pcp_mm', 'q_cms_obs'])
+    >>> df.unstack().shape
     (16436, 5)
     # get names of available static features
     >>> dataset.static_features
@@ -620,9 +622,9 @@ class CAMELS_AUS(_RainfallRunoff):
     ... # get names of available dynamic features
     >>> dataset.dynamic_features
     ... # get only selected dynamic features
-    >>> data = dataset.fetch(1, as_dataframe=True,
-    ...  dynamic_features=['airtemp_C_awap_max', 'pcp_mm_awap', 'et_morton_actual_SILO', 'q_cms_obs'])[1].unstack()
[…]
