### 2025-06-26 5c45615
added CAMELS_LUX
