+++ aqua_fetch/rr/_misc.py
@@ -1,7 +1,22 @@
 
-__all__ = ["DraixBleone"]
+__all__ = ["DraixBleone", "JialingRiverChina"]
 
+import os
+from typing import List
+
+import pandas as pd
+
+from .utils import check_attributes
 from .utils import _RainfallRunoff
+
+from ._map import (
+    observed_streamflow_cms, 
+    snow_depth, 
+    snow_water_equivalent,
+    mean_windspeed,
+    total_precipitation,
+    mean_dewpoint_temperature,
+    )
 
 
 class DraixBleone(_RainfallRunoff):
@@ -65,3 +80,129 @@ class DraixBleone(_RainfallRunoff):
         super().__init__(**kwargs)
 
         self._download()
+
+    def stations(self)->List[str]:
+        return ['BRU', 'LAV', 'MOU', 'ROU']
+
+    def _read_q_stn(self, stn:str):
+
+        fpath = os.path.join(self.path, f"DRAIXBLEONE_DRAIX_{stn}_DISCH.txt")
+        stn_df = pd.read_csv(fpath, sep=';', index_col=0, parse_dates=False,
+                             header=2, usecols=[0, 1, 2],
+                             )
+        
+        stn_df.index = pd.to_datetime(stn_df.index, format="%d/%m/%Y %H:%M:%S")
+
+        stn_df.rename(columns={'Valeur': observed_streamflow_cms()}, inplace=True)
+        stn_df.columns.name = 'dynamic_features'
+        stn_df.index.name = 'time'
+
+        # convert L/s to m3/s
+        # stn_df['runoff'] = stn_df['runoff'] * 1000.0
+
+        return stn_df
+
+    def _read_dynamic(self, stations, dynamic_features, st=None, en=None):
+        stations = check_attributes(stations, self.stations(), 'stations')
+
+        dyn = {}
+        for stn in stations:
+            
+            dyn[stn] = self._read_q_stn(stn)
+        return dyn
+
+    def _static_coords(self, stations = 'all'):
+        # from README.txt file
+        coords = {'BRU': (965694, 6345789, 801, 1.07, 87), 
+                  'LAV': (968818, 6343668, 850, 0.86, 32), 
+                  'MOU': (968688, 6343610, 847, 0.086, 46), 
+                  'ROU': (968828, 6343644, 852, 0.0013, 21)
+                  }
+        coords = pd.DataFrame.from_dict(coords, orient='index', 
+                                        columns=['long', 'lat', 'altitude', 'area', 'veg_cover_%'])
+        return coords
+    
+    def area(self, stations = 'all'):
+        area = {'BRU': 0.5, 'LAV': 0.5, 'MOU': 0.5, 'ROU': 0.5}
+        area = pd.DataFrame.from_dict(area, orient='index', columns=['area'])
+        return area
+
+
+class JialingRiverChina(_RainfallRunoff):
+    """
+    Dataset of 11 catchments in the upper, middle and lower reaches of the Jialing 
+    River mainstream basin, China . For more infromation on data see `Wang et al., 2024 <https://doi.org/10.1016/j.envsoft.2024.106091>`_.
+    The data consists of daily observations of weather variables and runoff from 
+    2010 to 2022. 
+
+    The dataset is available at `github link <https://github.com/AtrCheema/CVTGR-model>`_.
+    """
+    url = {
+        'Beibei.csv': 'https://raw.githubusercontent.com/AtrCheema/CVTGR-model/refs/heads/main/Data/OriginalData/Beibei.csv',
[…]
