+++ readme.md
@@ -155,7 +155,8 @@ mg_data_ohe.shape
 | CAMELS_DE      | 1555                   |                         | 21                       | 111                     | 1951 - 2020       | Germany                                     | [Loritz et al., 2024](https://essd.copernicus.org/preprints/essd-2024-318)                                  |
 | CAMELS_SE      | 50                     |                         | 4                        | 76                      | 1961 - 2020       | Sweden                                      | [Teutschbein et al., 2024](https://doi.org/10.1002/gdj3.239)                                                |
 | Caravan_DK     | 308                    |                         | 38                       | 211                     | 1981 - 2020       | Denmark                                     | [Koch, J. (2022)](https://doi.org/10.5281/zenodo.7962379)                                                   |
-| Caravan_NZ     | 369                    | 369                     | 5                        | 39                      | 1972 - 2024       | New Zealand                                 | [Bushra, et al., 2025](https://doi.org/10.5194/essd-2025-244)                                                   |
+| CAMELS_LUX     | 56                     | 56                      | 25                       | 56                      | 2004 - 2021       | Luxumbourg                                 | [Nijzink et al., 2025](https://doi.org/10.5194/essd-2024-482)                                                   |
+| CAMELS_NZ     | 369                    | 369                     | 5                        | 39                      | 1972 - 2024       | New Zealand                                 | [Bushra, et al., 2025](https://doi.org/10.5194/essd-2025-244)                                                   |
 | LamaHCE        | 859                    | 859                     | 22                       | 80                      | 1981 - 2019       | Central Europe                              | [Klingler et al., 2021](https://doi.org/10.5194/essd-13-4529-2021)                                          |
 | LamaHIce       | 111                    | 111                     | 36                       | 154                     | 1950 - 2021       | Iceland                                     | [Helgason and Nijssen 2024](https://doi.org/10.5194/essd-16-2741-2024)                                      |
 | HYSETS         | 14425                  |                         | 20                       | 30                      | 1950 - 2018       | North America                               | [Arsenault et al., 2020](https://doi.org/10.1038/s41597-020-00583-2)                                        |

[…]
