+++ aqua_fetch/rr/_camels.py
@@ -15,7 +15,7 @@ from .._project import utm_to_lat_lon
 from ..utils import get_cpus
 from ..utils import check_attributes, download, unzip
 
-from .._backend import netCDF4, xarray as xr
+from .._backend import netCDF4
 
 from ._map import (
     observed_streamflow_cms,
@@ -50,6 +50,8 @@ from ._map import (
     soil_moisture_layer2,
     soil_moisture_layer3,
     soil_moisture_layer4,
+    mean_dewpoint_temperature_at_2m,
+    catchment_area_with_specifier,
 )
 
 from ._map import (
@@ -269,72 +271,54 @@ class CAMELS_US(_RainfallRunoff):
 
         return stns
 
-    def _read_dynamic(
+    def _read_stn_dyn(
             self,
-            stations,
-            dynamic_features: Union[str, list] = 'all',
-            st=None,
-            en=None,
+            stn:str,
     ):
 
-        st, en = self._check_length(st, en)
-        dynamic_features = check_attributes(dynamic_features, self.dynamic_features, 'dynamic_features')
+        assert isinstance(stn, str)
+        df = None
+        dir_name = self.folders[self.data_source]
+        for cat in os.listdir(os.path.join(self.dataset_dir, dir_name)):
+            cat_dirs = os.listdir(os.path.join(self.dataset_dir, f'{dir_name}{SEP}{cat}'))
+            stn_file = f'{stn}_lump_cida_forcing_leap.txt'
+            if stn_file in cat_dirs:
+                df = pd.read_csv(os.path.join(self.dataset_dir,
+                                                f'{dir_name}{SEP}{cat}{SEP}{stn_file}'),
+                                    sep="\s+|;|:",
+                                    skiprows=4,
+                                    engine='python',
+                                    names=['Year', 'Mnth', 'Day', 'Hr', 'dayl(s)', 'prcp(mm/day)', 'srad(W/m2)',
+                                        'swe(mm)', 'tmax(C)', 'tmin(C)', 'vp(Pa)'],
+                                    )
+                df.index = pd.to_datetime(
+                    df['Year'].map(str) + '-' + df['Mnth'].map(str) + '-' + df['Day'].map(str))
 
-        dyn = {}
-        for station in stations:
+        flow_dir = os.path.join(self.dataset_dir, 'usgs_streamflow')
+        for cat in os.listdir(flow_dir):
+            cat_dirs = os.listdir(os.path.join(flow_dir, cat))
+            stn_file = f'{stn}_streamflow_qc.txt'
+            if stn_file in cat_dirs:
+                fpath = os.path.join(flow_dir, f'{cat}{SEP}{stn_file}')
+                q_df = pd.read_csv(fpath,
+                                    sep=r"\s+",
+                                    names=['station', 'Year', 'Month', 'Day', 'Flow', 'Flag'],
+                                    engine='python')
+                q_df.index = pd.to_datetime(
+                    q_df['Year'].map(str) + '-' + q_df['Month'].map(str) + '-' + q_df['Day'].map(str))
 
-            # attributes = check_attributes(dynamic_features, self.dynamic_features)
+        stn_df = pd.concat([
+            df[['dayl(s)', 'prcp(mm/day)', 'srad(W/m2)', 'swe(mm)', 'tmax(C)', 'tmin(C)', 'vp(Pa)']],
+            q_df['Flow']],
+            axis=1)
 
-            assert isinstance(station, str)
-            df = None
-            df1 = None
-            dir_name = self.folders[self.data_source]
-            for cat in os.listdir(os.path.join(self.dataset_dir, dir_name)):
-                cat_dirs = os.listdir(os.path.join(self.dataset_dir, f'{dir_name}{SEP}{cat}'))
[…]
