### 2025-06-27 d8a115f
definition of stations static_features and dynamic_features optional in child classes
