### 2025-06-27 hyex-research/AquaFetch#7 @AtrCheema
> ### Corrections
> * **Review Installation instructions**
>   I found the example uses `AquaFetch[all]` but the package metadata is `aqua‑fetch`. It should be:
>   pip install "aqua‑fetch[all] @ git+https://github.com/hyex-research/AquaFetch.git"
> * **Correct `fetch(...).unstack()` usage [example](https://aquafetch.readthedocs.io/en/latest/rr.html#aqua_fetch.rr.RainfallRunoff)**
>   Calling `.unstack()` directly on `dataset.fetch(..., as_dataframe=True)` raises an `AttributeError` because `fetch` returns a tuple (e.g., `(meta, df)`).
>
> ### Paper
> * **Correct typographical errors**
>
>   * **Location:** Section header on p. 2 (“Implementation and Archiecture”)
>   * Fix “Archiecture” → **Architecture**.
> * **Fix incorrect import name in example**
>
>   * **Location:** Code snippet (p. 3, lines 74–75)
>   * Change `RainfallRUnoff` → `RainfallRunoff`.
>   * Change `SWatch` → `SWatCh`.

Dear @jeiloh
The update instructions have been updated. The code in the examples and in particular the example you quoted has been corrected as well. The typos in the paper.md file have also been corrected.
