### 2025-07-19 f1d67c0
added color option in plot_functions
