### 2025-07-19 7527de1
correct use of area and coords for camels_fr
