### 2025-07-19 bdbb148
added camels_ch_chem
