+++ aqua_fetch/wq/__init__.py
@@ -38,3 +38,4 @@ from ._grimedb import GRiMeDB
 from ._misc import BuzzardsBay
 from ._misc import WhiteClayCreek
 from ._river_chem_siberia import RiverChemSiberia
+from ._camels_chem import Camels_Ch_Chem
