### 2025-07-23 a973839
use fiona for creating bndry_id_map dictionary and get_boundary returns list of rings
