### 2025-07-24 0fd80b7
added camels_fi dataset
