+++ aqua_fetch/rr/_camels.py
@@ -3,6 +3,7 @@ import json
 import glob
 import time
 import shutil
+import zipfile
 import warnings
 import concurrent.futures as cf
 from typing import Union, List, Dict
@@ -52,6 +53,8 @@ from ._map import (
     soil_moisture_layer4,
     mean_dewpoint_temperature_at_2m,
     catchment_area_with_specifier,
+    snow_water_equivalent_with_specifier,
+    snow_depth,
 )
 
 from ._map import (
@@ -62,9 +65,16 @@ from ._map import (
     gauge_elevation_meters,
     catchment_elevation_meters,
     urban_fraction,
+    urban_fraction_with_specifier,
     grass_fraction,
+    grass_fraction_with_specifier,
     crop_fraction,
+    crop_fraction_with_specifier,
     catchment_perimeter,
+    aridity_index,
+    med_catchment_elevation_meters,
+    soil_depth,
+    population_density,
     )
 
 # directory separator
@@ -3932,8 +3942,8 @@ class CAMELS_LUX(_RainfallRunoff):
         
         self._download(overwrite=overwrite)
 
-        if self.to_netcdf:
-            self._maybe_to_netcdf(f'camels_lux_dyn_{self.timestep}')
+        # if self.to_netcdf:
+        self._maybe_to_netcdf(f'camels_lux_dyn_{self.timestep}')
 
     @property
     def static_features(self) -> List[str]:
@@ -4124,3 +4134,264 @@ class CAMELS_ES(_RainfallRunoff):
     """
     """
     url = "https://zenodo.org/records/15040948"
+
+
+class CAMELS_FI(_RainfallRunoff):
+    """
+    Dataset of 320 Finnish catchments with 16 dynamic features and 106 static features.
+    The dynamic features span from 19610101 to 20231231 with daily timestep.
+    The data is downloaded from `Zenodo <https://zenodo.org/records/16257216>`_.
+
+    
+    Examples
+    ---------
+    >>> from aqua_fetch import CAMELS_FI
+    >>> dataset = CAMELS_FI()
+    >>> _, data = dataset.fetch(0.1, as_dataframe=True)
+    >>> data.shape
+    (368160, 32)   # 32 represents number of stations
+    Since data is a multi-index dataframe, we can get data of one station as below
+    >>> data['1156'].unstack().shape
+    (23010, 16)
+    If we don't set as_dataframe=True, then the returned data will be a xarray Dataset
+    >>> _, data = dataset.fetch(0.1)
+    >>> type(data)
+        xarray.core.dataset.Dataset
+    >>> data.dims
+    FrozenMappingWarningOnValuesAccess({'time': 23010, 'dynamic_features': 16})
+    >>> len(data.data_vars)
+        32
+    >>> _, df = dataset.fetch(stations=1, as_dataframe=True)  # get data of only one random station
+    >>> df = df.unstack() # the returned dataframe is a multi-indexed dataframe so we have to unstack it
+    >>> df.shape
+    (23010, 16)
+    # get name of all stations as list
+    >>> stns = dataset.stations()
+    >>> len(stns)
+    320
+    # get data by station id
+    >>> _, df = dataset.fetch(stations='1156', as_dataframe=True)
+    >>> df.unstack().shape
+    (23010, 16)
+    # get names of available dynamic features
+    >>> dataset.dynamic_features
+    # get only selected dynamic features
+    >>> _, df = dataset.fetch(1, as_dataframe=True,
+    ... dynamic_features=['pcp_mm', 'snowdepth_m', 'airtemp_C_mean', 'pet_mm', 'q_cms_obs'])
+    >>> df.unstack().shape
+    (23010, 5)
+    # get names of available static features
[…]
