### 2025-07-27 6a66cf3
added Slovenia dataset
