+++ aqua_fetch/rr/_estreams.py
@@ -7,12 +7,15 @@ __all__ = [
     "Italy", 
     "Poland",
     "Portugal",
+    "Slovenia"
     ]
 
 import os
 import time
 import warnings
+import urllib.parse
 from io import StringIO
+from datetime import datetime
 import concurrent.futures as cf
 from urllib.error import HTTPError
 from typing import Union, List, Dict
@@ -54,6 +57,8 @@ from ._map import (
     observed_streamflow_cms,
     )
 
+
+# todo : a lot of methods in subclasses of _EStreams are redundant
 
 class EStreams(_RainfallRunoff):
     """
@@ -494,6 +499,27 @@ class _EStreams(_RainfallRunoff):
     def end(self) -> pd.Timestamp:
         return pd.Timestamp('2023-06-30')
 
+    def stations(self) -> List[str]:
+        """
+        Returns a list of all station names. Note that the `basin_id` column is 
+        used as the station name.
+        """
+        return self._stations
+    
+    def gauge_id_basin_id_map(self)->dict:
+        """
+        For example for Portugal, it is
+        guage_id : '03J/02H'
+        basin_id 'PT000001'
+        '03J/02H' -> 'PT000001'
+
+        for Slovenia, it is
+        gauge id : 1060
+        basin_id : SI000001
+        '1060' -> 'SI000001'
+        """
+        return {k:v for v,k in self.md['gauge_id'].to_dict().items()}
+
     def _fetch_dynamic_features(
             self,
             stations: list,
@@ -686,11 +712,11 @@ class Finland(_EStreams):
     Data of 669 catchments of Finland. 
     The observed streamflow data is downloaded from 
     https://wwwi3.ymparisto.fi .
-    The meteorological data, static catchment 
+    The meteorological data, stattic catchment 
     features and catchment boundaries are
     taken from :py:class:`aqua_fetch.EStreams` follwoing the works
     of `Nascimento et al., 2024 <https://doi.org/10.5194/hess-25-471-2021>`_ . Therefore,
-    the number of staic features are 214 and dynamic features are 10 and the
+    the number of static features are 214 and dynamic features are 10 and the
     data is available from 2012-01-01 to 2023-06-30.
     """
     def __init__(
@@ -960,7 +986,7 @@ class Ireland(_EStreams):
     features and catchment boundaries are
     taken from :py:class:`aqua_fetch.EStreams` follwoing the works
     of `Nascimento et al., 2024 <https://doi.org/10.5194/hess-25-471-2021>`_ project. Therefore,
-    the number of staic features are 214 and dynamic features are 10 and the
+    the number of static features are 214 and dynamic features are 10 and the
     data is available from 1992-01-01 to 2020-06-31.
     """
     def __init__(
@@ -1345,7 +1371,7 @@ class Italy(_EStreams):
     features and catchment boundaries are
     taken from :py:class:`aqua_fetch.EStreams` follwoing the works
     of `Nascimento et al., 2024 <https://doi.org/10.5194/hess-25-471-2021>`_ . Therefore,
-    the number of staic features are 214 and dynamic features are 10 and the
+    the number of static features are 214 and dynamic features are 10 and the
     data is available from 1992-01-01 to 2020-06-31.
     """
     def __init__(
@@ -1473,7 +1499,7 @@ class Poland(_EStreams):
     features and catchment boundaries are
     taken from :py:class:`aqua_fetch.EStreams` follwoing the works
     of `Nascimento et al., 2024 <https://doi.org/10.5194/hess-25-471-2021>`_ . Therefore,
[…]
