+++ aqua_fetch/rr/_estreams.py
@@ -718,6 +718,48 @@ class Finland(_EStreams):
     of `Nascimento et al., 2024 <https://doi.org/10.5194/hess-25-471-2021>`_ . Therefore,
     the number of static features are 214 and dynamic features are 10 and the
     data is available from 2012-01-01 to 2023-06-30.
+
+    Examples
+    ---------
+    >>> from aqua_fetch import Finland
+    >>> dataset = Finland()
+    >>> _, data = dataset.fetch(0.1)  # the returned data will be a xarray Dataset
+    >>> type(data)
+        xarray.core.dataset.Dataset
+    >>> data.dims
+    FrozenMappingWarningOnValuesAccess({'time': 4199, 'dynamic_features': 10})
+    >>> len(data.data_vars)  # number of stations for which data has been fetched
+        66
+    >>> _, data = dataset.fetch(stations=1)  # get data of only one random station
+    # get name of all stations as list
+    >>> stns = dataset.stations()
+    >>> len(stns)
+    669
+    # get data by station id
+    >>> _, data = dataset.fetch(stations='FI000001')
+    # get names of available dynamic features
+    >>> dataset.dynamic_features
+    # get only selected dynamic features
+    >>> _, data = dataset.fetch(1,
+    ... dynamic_features=['pcp_mm', 'rh_%', 'airtemp_C_mean', 'pet_mm', 'q_cms_obs'])
+    # get names of available static features
+    >>> dataset.static_features
+    # get data of 10 random stations
+    >>> _, data = dataset.fetch(10)
+    >>> len(data.data_vars)
+    10
+    # If we want to get both static and dynamic data
+    >>> static, dynamic = dataset.fetch(stations='FI000001', static_features="all")
+    >>> static.shape, len(dynamic.data_vars)
+    ((1, 214), 1)
+    >>> coords = dataset.stn_coords() # returns coordinates of all stations
+    >>> coords.shape
+        (669, 2)
+    >>> dataset.stn_coords('FI000001')  # returns coordinates of station whose id is FI000001
+        64.226288	27.736528
+    >>> dataset.stn_coords(['FI000001', 'FI000002'])  # returns coordinates of two stations
+    FI000001	64.226288	27.736528
+    FI000002	64.226288	27.736528
     """
     def __init__(
             self, 
@@ -735,10 +777,6 @@ class Finland(_EStreams):
     @property
     def start(self)->pd.Timestamp:
         return pd.Timestamp('2012-01-01')
-
-    def stations(self)->List[str]:
-        """returns the `basin_id` of the stations"""
-        return self._stations
 
     def gauge_id_basin_id_map(self)->dict:
         # guage_id '5902650'
@@ -988,6 +1026,48 @@ class Ireland(_EStreams):
     of `Nascimento et al., 2024 <https://doi.org/10.5194/hess-25-471-2021>`_ project. Therefore,
     the number of static features are 214 and dynamic features are 10 and the
     data is available from 1992-01-01 to 2020-06-31.
+
+    Examples
+    ---------
+    >>> from aqua_fetch import Ireland
+    >>> dataset = Ireland()
+    >>> _, data = dataset.fetch(0.1)  # the returned data will be a xarray Dataset
+    >>> type(data)
+        xarray.core.dataset.Dataset
+    >>> data.dims
+    FrozenMappingWarningOnValuesAccess({'time': 26844, 'dynamic_features': 10})
+    >>> len(data.data_vars)  # number of stations for which data has been fetched
+        46
+    >>> _, data = dataset.fetch(stations=1)  # get data of only one random station
+    # get name of all stations as list
+    >>> stns = dataset.stations()
+    >>> len(stns)
[…]
