+++ aqua_fetch/wq/__init__.py
@@ -39,3 +39,4 @@ from ._misc import BuzzardsBay
 from ._misc import WhiteClayCreek
 from ._river_chem_siberia import RiverChemSiberia
 from ._camels_chem import Camels_Ch_Chem
+from ._oligotrend import Oligotrend
