+++ aqua_fetch/__init__.py
@@ -99,6 +99,7 @@ from .wq import BuzzardsBay
 from .wq import WhiteClayCreek
 from .wq import RiverChemSiberia
 from .wq import Camels_Ch_Chem
+from .wq import Oligotrend
 
 # *** Miscellaneous ***
 
@@ -199,6 +200,7 @@ ALL_DATASETS = [
     WhiteClayCreek.__class__.__name__,
     RiverChemSiberia.__class__.__name__,
     Camels_Ch_Chem.__class__.__name__,
+    Oligotrend.__class__.__name__,
     
     ec_removal_biochar.__name__,
     cr_removal.__name__,

[…]
