### 2025-07-28 0b96b7d
updated minor bugs and docs
