+++ tests/rr/test_misc.py
@@ -1,7 +1,7 @@
 
 import os
 import site
-wd_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+wd_dir = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 site.addsitedir(wd_dir)
 
 import logging
@@ -9,108 +9,36 @@ import logging
 if __name__ == "__main__":
     logging.basicConfig(filename='test_estreams.log', filemode='w', level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 
-import pandas as pd
+from aqua_fetch.rr import DraixBleone
+from aqua_fetch.rr import JialingRiverChina
 
-from aqua_fetch.rr import EStreams
-
-from utils import (
-    test_coords,
-    test_stations,
-    test_area,
-    test_boundary,
-    test_plot_stations,
-    test_fetch_static_feature,
-    test_fetch_dynamic_features
-    )
-
-gscad_path = '/mnt/datawaha/hyex/atr/gscad_database/raw'
-
-ds = EStreams(path=gscad_path, verbosity=3)
-
-assert ds.md.shape == (15047, 24), ds.md.shape
-
-assert len(ds.countries) == 39, len(ds.agencies)
-
-assert len(ds.country_stations('IT')) == 767, len(ds.country_stations('IT'))
-
-assert len(ds.country_stations('ES')) == 1440, len(ds.country_stations('ES'))
-
-assert len(ds.country_stations('IE')) == 464, len(ds.country_stations('IE'))
-
-assert len(ds.country_stations('PL')) == 1287, len(ds.country_stations('PL'))
+from utils import test_fetch_dynamic_features
 
 
-def test_coords_():
+def test_jilingriverchina():
+    dataset = JialingRiverChina()
 
-    test_coords(ds)
+    assert len(dataset.stations()) == 11
 
-    assert ds.stn_coords(countries='IE').shape == (464, 2), ds.stn_coords(countries='IE').shape
+    for stn in dataset.stations():
+        df = dataset._read_stn_dyn(stn)
+        assert len(df.shape) == 2
 
-    assert ds.stn_coords(stations='IEEP0281').shape == (1, 2), ds.stn_coords(stations='IEEP0281').shape
+    assert len(dataset.dynamic_features) == 43
+
+    test_fetch_dynamic_features(dataset, dataset.stations()[0], as_dataframe=True)
 
     return
 
 
-def test_area_():
+ds = DraixBleone(path='/mnt/datawaha/hyex/atr/data', verbosity=4)
 
-    test_area(ds)
+assert len(ds.stations()) == 4
 
-    assert ds.area(countries='IE').shape == (464, ), ds.area(countries='IE').shape
+assert len(ds.static_features) == 5
 
-    assert ds.area(stations='IEEP0281').shape == (1, ), ds.area(stations='IEEP0281').shape
+for stn in ds.stations():
+    df = ds._read_stn_dyn(stn)
+    assert len(df.shape) == 2
 
-    return
-
-
-def test_fetch_static_features():
-    out = ds.fetch_static_features(countries='IE')
-    assert out.shape == (464, 208), out.shape
-
-    test_fetch_static_feature(ds, 'IEEP0281', 15047, 208)
-    return
-
-
-def test_meteo():
-    out = ds.meteo_data_station('IEEP0281')
-
-    assert isinstance(out, pd.DataFrame), type(out)
-
-    assert out.shape[1] == 9, out.shape
-
-    out = ds.meteo_data(countries='IE')
-    assert len(out) == 464, len(out)
-
-    out = ds.meteo_data()
[…]
