### 2025-07-28 97614bf
get_boundary return fiona geometry instead of list of arrays
