+++ aqua_fetch/_datasets.py
@@ -192,54 +192,18 @@
 # https://www.tr32db.uni-koeln.de/search/view.php?dataID=1837
 # https://www.tr32db.uni-koeln.de/search/view.php?dataID=1760
 # https://www.tr32db.uni-koeln.de/search/view.php?dataID=1761
-
 import os
-import glob
-import random
-import warnings
-from typing import Union, List
 
 import numpy as np
-import pandas as pd
-
-from ._backend import netCDF4, xarray as xr
 
 from .download_pangaea import PanDataSet
-from .utils import maybe_download, download_and_unzip, unzip_all_in_dir, download
-from .utils import check_attributes
+from .utils import maybe_download, download_and_unzip
 
 
 SEP = os.sep
 # TODO, add visualization
 
 # TODO all available datasets should be available using a single interface instead of importing each separately
-DATASETS = [
-    'ISWDC',
-    'SEBAL_ET_CHINA',
-    'GeoChemMatane',
-    'PrecipBerlin',
-    'HydroChemJava',
-    'WaterChemVictoriaLakes',
-    'WaterChemEcuador',
-    'HydrocarbonsGabes',
-    'SedimentAmersee',
-    'FlowTetRiver',
-    'HoloceneTemp',
-    'RiverTempEroo',
-    'StreamTempSpain',
-    'FlowSedDenmark',
-    'FlowSamoylov',
-    'EtpPcpSamoylov',
-    'RiverIsotope',
-    'WQCantareira',
-    'RiverTempSpain',
-    'HydrometricParana',
-    'FlowBenin',
-    'YamaguchiClimateJp',
-    'WQJordan2',
-    'WQJordan',
-    'Weisssee'
-            ]
 
 
 class Datasets(object):
@@ -395,616 +359,3 @@ class Datasets(object):
         if self.remove_zip:
             self.remove_zip_files()
         return
-
-
-class Weisssee(Datasets):
-
-    dynamic_attributes = ['Precipitation_measurements',
-                          'long_wave_upward_radiation',
-                          'snow_density_at_30cm',
-                          'long_wave_downward_radiation'
-                          ]
-
-    url = '10.1594/PANGAEA.898217'
-
-    def __init__(self, path=None, overwrite=False, **kwargs):
-        super(Weisssee, self).__init__(path=path, **kwargs)
-        #self.path = path
-        self.download_from_pangaea(overwrite=overwrite)
-
-    def fetch(self, **kwargs):
-        """
-        Examples
-        --------
-            >>> from aqua_fetch import Weisssee
-            >>> dataset = Weisssee()
-            >>> data = dataset.fetch()
-        """
-
-        data = {}
-        for f in self.data_files:
-            fpath = os.path.join(self.path, f)
-            df = pd.read_csv(fpath, **kwargs)
-
-            if 'index_col' in kwargs:
-                df.index = pd.to_datetime(df.index)
-
-            data[f.split('.txt')[0]] = df
-
-        return data
-
-
-class ETP_CHN_SEBAL(Datasets):
-
-    url = "https://zenodo.org/record/4218413#.YBNhThZS-Ul"
-
-
-class ISWDC(Datasets):
-
-    url = "https://zenodo.org/record/2616035#.YBNl5hZS-Uk"
-
-
-class WQJordan(Weisssee):
-    """Jordan River water quality data of 9 variables for two variables."""
-    url = 'https://doi.pangaea.de/10.1594/PANGAEA.919103'
-
-
-class WQJordan2(Weisssee):
-    """Stage and Turbidity data of Jordan River"""
-    url = '10.1594/PANGAEA.919104'
-
-
[…]
