### 2025-07-29 4a1da0c
removed shapely and pyshp as optional dependencies
