+++ paper/paper.bib
@@ -422,3 +422,11 @@ DOI = {10.5194/essd-12-2459-2020}
   doi          = {10.5281/zenodo.7962379},
   url          = {https://doi.org/10.5281/zenodo.7962379},
 }
+
+
+@software{fiona,
+author = {Sean, Gillies and Rene, Buffat and Joshua, Arnott and Mike W, Taves and Kevin, Wurster and Alan D, Snow and Micah, Cochran and Others, Many},
+license = {BSD-3-Clause},
+title = {{Fiona}},
+url = {https://github.com/Toblerity/Fiona}
+}
