### 2025-07-29 dcc279c
updated examples on readme
