+++ readme.md
@@ -54,7 +54,7 @@ from aqua_fetch import RainfallRunoff
 dataset = RainfallRunoff('CAMELS_AUS')  # instead of CAMELS_AUS, you can provide any other dataset name
 _, df = dataset.fetch(stations=1, as_dataframe=True)
 df = df.unstack() # the returned dataframe is a multi-indexed dataframe so we have to unstack it
-df.columns = df.columns.get_level_values('dynamic_features')
+df.columns = df.columns.levels[1]
 df.shape
    (26388, 28)
 # get name of all stations as list
@@ -63,38 +63,40 @@ len(stns)
    561
 # get data of 10 % of stations as dataframe
 _, df = dataset.fetch(0.1, as_dataframe=True)
-df.shape
-   (550784, 28)
+df.shape  # (738864, 56)
+
 # The returned dataframe is a multi-indexed data
-df.index.names == ['time', 'dynamic_features']
-    True
+df.index.names   # ['time', 'dynamic_features'] 
+
 # get data by station id
 _, df = dataset.fetch(stations='912101A', as_dataframe=True)
-df.unstack().shape
-    (26388, 28)
+df.unstack().shape  # (26388, 28)
+
 # get names of available dynamic features
 dataset.dynamic_features
+
 # get only selected dynamic features
 _, data = dataset.fetch(1, as_dataframe=True,
-...  dynamic_features=['airtemp_C_mean_awap', 'pcp_mm_awap', 'aet_mm_silo_morton', 'q_cms_obs'])
-data.unstack().shape
-   (26388, 4)
+...  dynamic_features=['airtemp_C_mean_agcd', 'pcp_mm_agcd', 'aet_mm_silo_morton', 'q_cms_obs'])
+data.unstack().shape  # (26388, 4)
+
 # get names of available static features
 dataset.static_features
+
 # get data of 10 random stations
 df = dataset.fetch(10, as_dataframe=True)
-df.shape  # remember this is a multiindexed dataframe
-   (26388, 280)
+df.shape  # remember this is a multiindexed dataframe  with shape (26388, 280)
 # If we get both static and dynamic data
 static, dynamic = dataset.fetch(stations='912101A', static_features="all", as_dataframe=True)
-static.shape, dynamic.unstack().shape
-((1, 187), (26388, 28))
-coords = dataset.stn_coords() # returns coordinates of all stations
-coords.shape
-    (561, 2)
-dataset.stn_coords('912101A')  # returns coordinates of station whose id is 912101A
-    -18.643612	139.253052
-dataset.stn_coords(['912101A', '912105A'])  # returns coordinates of two stations
+static.shape, dynamic.unstack().shape   # ((1, 187), (26388, 28))
+
+# get coordinates of all stations
+coords = dataset.stn_coords()
+coords.shape  #     (561, 2)
+# get coordinates of station whose id is 912101A
+dataset.stn_coords('912101A')       # -18.643612	139.253052
+# get coordinates of two stations
+dataset.stn_coords(['912101A', '912105A'])
 ```
 
 The datasets related to surface water quality are available using functional or objected-oriented API
@@ -104,16 +106,16 @@ quality related datasets. For complete name of Python functions and classes see 
 ```python
 from aqua_fetch import busan_beach
 dataframe = busan_beach()
-dataframe.shape
-(1446, 14)
+dataframe.shape  # (1446, 14)
+
 dataframe = busan_beach(target=['tetx_coppml', 'sul1_coppml'])
-dataframe.shape
-(1446, 15)
+dataframe.shape  # (1446, 15)
 
[…]
