### 2025-07-29 55ed32c
updated bib info
