+++ paper/paper.bib
@@ -343,6 +343,8 @@ author = {{NSF Unidata} and Davis, Glenn and Rew, Russ and Heimbigner, Dennis an
 doi = {10.5065/D6H70CW6},
 license = {BSD-3-Clause},
 title = {{NetCDF-C }},
+year = {2024},
+version = {1.7.2},
 url = {https://github.com/Unidata/netcdf-c}
 }
 
@@ -428,5 +430,7 @@ DOI = {10.5194/essd-12-2459-2020}
 author = {Sean, Gillies and Rene, Buffat and Joshua, Arnott and Mike W, Taves and Kevin, Wurster and Alan D, Snow and Micah, Cochran and Others, Many},
 license = {BSD-3-Clause},
 title = {{Fiona}},
+year = {2024},
+version = {1.10.0},
 url = {https://github.com/Toblerity/Fiona}
 }
