+++ setup.py
@@ -57,7 +57,6 @@ setup(
 
     package_data={'aqua_fetch': ['data/portugal_stn_codes.csv']},
     include_package_data=True,
-    #data_files=[('data', ['data/portugal_stn_codes.csv'])],
 
     classifiers=[
         'Development Status :: 4 - Beta',
