+++ CONTRIBUTING.rst
@@ -26,12 +26,12 @@ rainfall-runoff
 Consider that you want to add a new rainfall-runoff dataset named ``NewDataSource``.
 
 - create a new file in ``aqua_fetch/rr/_newdatasource.py`` file
-- create a new class in this file which must inherit from ``_RainfallRunoff`` class
+- create a new class in this file which must inherit from `_RainfallRunoff <https://aquafetch.readthedocs.io/en/latest/rr.html#aqua_fetch.rr._RainfallRunoff>`_ class
 - implement the following methods. The signature of these methods/attributes must be same as the parent class.
 
     - ``__init__``  -> within this method, the logic to download the data must be implemented. The data must be downloaded within `path`
-    - ``_read_stn_dyn`` -> must return a pandas DataFrame with with columns of time series of /dynamic/hydrometeorological data for a single station and index as time
-    - ``_read_static``  -> must return a pandas DataFrame with columns as names of static_features and index as station ids
+    - ``_read_stn_dyn`` -> It defines the logic to read dynamic/time series data for a single station and it must return a pandas DataFrame with with columns of time series of /dynamic/hydrometeorological data for a single station and index as time
+    - ``_read_static``  -> It defines the logic to read static data for all stations and it must return a pandas DataFrame with columns as names of static_features and index as station ids
 - The implementation/creation of following methods/attributes is optional but recommended
 
     - ``dynamic_features`` -> must return a list of dynamic features that are available in the dataset, if not implemented then the column names of _read_stn_dyn() will be used
@@ -41,7 +41,7 @@ Consider that you want to add a new rainfall-runoff dataset named ``NewDataSourc
     - ``start`` -> return the start date of the dataset
     - ``stn_coords``
     - ``q_mmd``
-    - ``boundary_file``  -> the path to the boundary file for the dataset.
+    - ``boundary_file``  -> the path to the file containing catchment boudaries for the dataset.
 - import this class in ``aqua_fetch/__init__.py`` file in aqua_fetch folder and change the ``ALL_DATASETS`` lists to include this new class
 - update the ``DATASETS`` dictionary in ``aqua_fetch/rr/__init__.py`` file to include this new class
 - Write tests for this new data source in tests directory where you compare the files generated by this new data source with the raw files.
