### 2025-07-29 #8051 @AtrCheema
@ethanwhite @tamnva @jeiloh
Thank you very much for your time, constructive suggestions to improve the repository, and for identifying bugs in the code.

We have addressed the issues raised by both @tamnva and @jeiloh and have published a new release candidate ([1.0.0.rc3](https://pypi.org/project/aqua-fetch/1.0.0rc3/)) on PyPI. This means that running ``pip install aqua-fetch`` now installs the most up-to-date version of the package. The documentation on readthedocs has also been updated. Since most of the issues identified by the reviewers were due to errors in the example code within docstrings—and those have now been fixed by updating the examples—I will focus below on addressing the remaining comments and suggestions.

### Project Name
The project/repository is named ``AquaFetch``, while the source code resides in the ``aqua_fetch`` package. This distinction was intentional and follows a common convention in the Python ecosystem. For example, the well-known machine learning library  [`scikit-learn`](https://github.com/scikit-learn/scikit-learn) is imported as[``sklearn``](https://github.com/scikit-learn/scikit-learn/tree/main/sklearn). On PyPI, our package is automatically listed as aqua-fetch, as underscores (``_``) are typically converted to hyphens (``-``). This behavior is also seen in other packages, such as the easy_mpl project on [GitHub](https://github.com/Sara-Iftikhar/easy_mpl) and [PyPi](https://pypi.org/project/easy_mpl).

### Installation with Minimal Dependencies
There were some import errors when the package was installed with minimal dependencies. We have resolved these issues. Now, even if matplotlib or xarray are not installed, most core functionalities of the package remain available and functional.

### Author Contributions
There are three authors of the paper. I, @AtrCheema, am the submitting author and have made the majority of the commits. However, commit frequency does not necessarily reflect the full extent of individual contributions. @hylken was involved in API design, overall project supervision, paper writing and editing, bug-fixing suggestions, and also contributed to securing funding. @Sara-Iftikhar was primarily responsible for maintaining the ``wwt`` submodule.

### GitHub Actions Testing
We consciously avoided automatic testing with GitHub Actions because the test suite requires downloading large datasets, which is time-consuming. Moreover, the free GitHub Actions tier has limited compute resources. However, we manually run all tests in the tests/ directory before every new release. These tests provide comprehensive coverage of the codebase. Additionally,  [some code examples](https://aquafetch.readthedocs.io/en/latest/auto_examples/index.html) are executed during documentation [builds](https://app.readthedocs.org/projects/aquafetch/builds/) on ReadTheDocs, which, although limited, serve as another layer of testing.

### Contributing Guidelines
We have updated the [CONTRIBUTING.rst](https://github.com/hyex-research/AquaFetch/blob/master/CONTRIBUTING.rst) file with detailed instructions on how to add a new dataset. Since the ``rr`` submodule is currently the most mature, its contribution guidelines are more comprehensive. Adding a new dataset generally involves creating a new class that inherits from the [``_RainfallRunoff``](https://aquafetch.readthedocs.io/en/latest/rr.html#aqua_fetch.rr._RainfallRunoff) base class and overriding three methods: ``__init__``, ``_read_stn_dyn``, and ``_read_static``.

### Unique Value of the Project
We acknowledge that the datasetzoo submodule in the NeuralHydrology project is similar in spirit to our ``rr`` submodule. While the reviewers rightly point out that AquaFetch includes a significantly larger number of rainfall-runoff datasets ([36 in AquaFetch](https://aquafetch.readthedocs.io/en/latest/rr.html#list-of-datasets) versus 8 in NeuralHydrology), the key novelty of our project lies in its diversity. Unlike NeuralHydrology, which focuses solely on rainfall-runoff datasets, AquaFetch spans at least three distinct domains within water resources, thereby catering to a broader community of scientists and developers.

### Other Major Changes
- We added several new datasets in ``rr`` and ``wq`` submodules. Figures, tables, and numerical summaries in the paper.md, docstrings, and README.md have all been updated accordingly.
- The project’s dependency on pyshp and shapely has been removed. Spatial data is now read using the [fiona](https://fiona.readthedocs.io/) package, which supports a wider variety of spatial formats (e.g., pyshp does not support .gpkg files).
