### 2025-08-06 hyex-research/AquaFetch#8 @tamnva
Just one small comment. The function `dataset.fetch()` in the readme example requires very high computational resources (e.g., with the CAMELS_AUS data) eventhough users just want to extract a single features. It would be nice if the developers can optimize this
