### 2025-08-06 #8051 @jeiloh
This looks good to me as well.

One quick note: the current example uses CAMELS_AUS, which took about 45 minutes for me to download. You might consider switching to a lighter dataset for the example, though I leave this to your discretion.

Thank you again for addressing my comments—excellent work!
