### 2025-08-06 hyex-research/AquaFetch#8 @tamnva
Hi, in both cases (when I extracting single dynamic feature and extracting data for a single stations?) - the code I took from your usage section in the README (it takes around half a minutes with my PC). I have xarray installed already
