### 2025-08-07 e222894
updated CAMELS_SE and use it for readme
