+++ aqua_fetch/rr/_camels.py
@@ -13,10 +13,10 @@ import pandas as pd
 
 from .utils import _RainfallRunoff
 from .._geom_utils import utm_to_lat_lon
-from ..utils import get_cpus
+from ..utils import get_cpus, download_and_unzip
 from ..utils import check_attributes, download, unzip
 
-from .._backend import netCDF4
+from .._backend import netCDF4, xarray as xr
 
 from ._map import (
     observed_streamflow_cms,
@@ -453,7 +453,7 @@ class CAMELS_GB(_RainfallRunoff):
         self._maybe_to_netcdf('camels_gb_dyn')
 
         if not os.path.exists(self.boundary_file):
-            unzip(self.data_path)
+            unzip(self.data_path, verbosity=self.verbosity)
 
     @property
     def boundary_file(self) -> os.PathLike:
@@ -747,12 +747,12 @@ class CAMELS_AUS(_RainfallRunoff):
             if os.path.exists(fpath) and overwrite:
                 os.remove(fpath)
                 if verbosity > 0: print(f"Re-downloading {_file} from {url + _file} at {fpath}")
-                download(url + _file, outdir=self.path, fname=_file)
+                download(url + _file, outdir=self.path, fname=_file, verbosity=self.verbosity)
 
             elif not os.path.exists(fpath):
                 if verbosity > 0:
                     print(f"Downloading {_file} from {url + _file} at {fpath}")
-                download(url + _file, outdir=self.path, fname=_file)
+                download(url + _file, outdir=self.path, fname=_file, verbosity=self.verbosity)
             elif verbosity > 0:
                 print(f"{_file} already exists at {self.path}")
 
@@ -2014,10 +2014,10 @@ class CAMELS_SE(_RainfallRunoff):
     """
 
     url = {
-        'catchment properties.zip': "https://snd.se/sv/catalogue/download/dataset/2023-173/1?principal=user.uu.se&filename=catchment+properties.zip",
-        'catchment time series.zip': 'https://snd.se/sv/catalogue/download/dataset/2023-173/1?principal=user.uu.se&filename=catchment+time+series.zip',
-        'catchment_GIS_shapefiles.zip': "https://snd.se/sv/catalogue/download/dataset/2023-173/1?principal=user.uu.se&filename=catchment_GIS_shapefiles.zip",
-        'Documentation_2024-01-02.pdf': 'https://snd.se/en/catalogue/download/dataset/2023-173/1?principal=user.uu.se&filename=Documentation_2024-01-02.pdf&isDoc=y'
+ 'catchment properties.zip': 'https://api.researchdata.se/dataset/2023-173/1/file/data?filePath=catchment+properties.zip',
+ 'catchment time series.zip': 'https://api.researchdata.se/dataset/2023-173/1/file/data?filePath=catchment+time+series.zip',
+ 'catchment_GIS_shapefiles.zip': 'https://api.researchdata.se/dataset/2023-173/1/file/data?filePath=catchment_GIS_shapefiles.zip',
+ 'Documentation_2024-01-02.pdf': 'https://api.researchdata.se/dataset/2023-173/1/file/documentation?filePath=Documentation_2024-01-02.pdf'
     }
 
     def __init__(
@@ -2045,8 +2045,8 @@ class CAMELS_SE(_RainfallRunoff):
             if not os.path.exists(fpath) and not overwrite:
                 if verbosity > 0:
                     print(f"Downloading {_file} from {url}")
-                download(url, outdir=self.path, fname=_file, )
-                unzip(self.path)
+                download(url, outdir=self.path, fname=_file, verbosity=self.verbosity)
+                unzip(self.path, verbosity=self.verbosity)
             else:
                 if self.verbosity > 0: print(f"{_file} at {self.path} already exists")
 
[…]
