+++ docs/source/quickstart.rst
@@ -16,50 +16,55 @@ however, the same interface can be used to access `all other datasets <https://a
 .. code-block:: python
 
     >>> from aqua_fetch import RainfallRunoff
-    >>> dataset = RainfallRunoff('CAMELS_AUS')  # instead of CAMELS_AUS, you can provide any other dataset name
+    >>> dataset = RainfallRunoff('CAMELS_SE')  # instead of CAMELS_SE, you can provide any other dataset name
     >>> _, df = dataset.fetch(stations=1, as_dataframe=True)
     >>> df = df.unstack() # the returned dataframe is a multi-indexed dataframe so we have to unstack it
     >>> df.columns = df.columns.levels[1]
     >>> df.shape
-       (26388, 28)
+    (21915, 4)
+    ...
     ... # get name of all stations as list
     >>> stns = dataset.stations()
     >>> len(stns)
-       561
+       50
     ... # get data of 10 % of stations as dataframe
     >>> _, df = dataset.fetch(0.1, as_dataframe=True)
     >>> df.shape
-       (738864, 56)
+       (87660, 5)
+    ...
     ... # The returned dataframe is a multi-indexed data
     >>> df.index.names
         ['time', 'dynamic_features']
     ... # get data by station id
-    >>> _, df = dataset.fetch(stations='912101A', as_dataframe=True).unstack()
+    >>> _, df = dataset.fetch(stations='5', as_dataframe=True).unstack()
     >>> df.shape
-        (26388, 28)
+        (21915, 4)
     ... # get names of available dynamic features
     >>> dataset.dynamic_features
     ... # get only selected dynamic features
     >>> _, data = dataset.fetch(1, as_dataframe=True,
-    ...  dynamic_features=['airtemp_C_mean_agcd', 'pcp_mm_agcd', 'aet_mm_silo_morton', 'q_cms_obs']).unstack()
+    ...  dynamic_features=['pcp_mm', 'airtemp_C_mean', 'q_cms_obs']).unstack()
     >>> data.shape
-       (26388, 4)
+       (21915, 3)
+    ...
     ... # get names of available static features
     >>> dataset.static_features
     ... # get data of 10 random stations
     >>> _, df = dataset.fetch(10, as_dataframe=True)
     >>> df.shape  # remember this is a multiindexed dataframe
-       (26388, 280)
+       (87660, 10)
+    ...
     # If we get both static and dynamic data
-    >>> static, dynamic = dataset.fetch(stations='912101A', static_features="all", as_dataframe=True)
+    >>> static, dynamic = dataset.fetch(stations='5', static_features="all", as_dataframe=True)
     >>> static.shape, dynamic.shape
-    ((1, 187), (26388, 28))
+    ((1, 76), (21915, 4))
+    ...
     >>> coords = dataset.stn_coords() # returns coordinates of all stations
     >>> coords.shape
-        (561, 2)
-    >>> dataset.stn_coords('912101A')  # returns coordinates of station whose id is 912101A
-        18.3861	80.3917
-    >>> dataset.stn_coords(['912101A', '912105A'])  # returns coordinates of two stations
+        (50, 2)
+    >>> dataset.stn_coords('5')  # returns coordinates of station whose id is 5
+        68.035599	21.9758
+    >>> dataset.stn_coords(['5', '736'])  # returns coordinates of two stations
 
 
 Water Quality Datasets

[…]
