### 2025-08-07 480c012
modified sentence in paper
