### 2025-08-11 0349737
faster io operations for camels_aus
