### 2025-08-15 70c50e7
return dict of df instead of multiindex df for rr
