+++ aqua_fetch/rr/_hype.py
@@ -18,35 +18,76 @@ class HYPE(_RainfallRunoff):
 
     Examples
     --------
-        >>> from aqua_fetch import HYPE
-        >>> dataset = HYPE()
-        ... # get data of 5% of stations
-        >>> df = dataset.fetch(stations=0.05, as_dataframe=True)  # returns a multiindex dataframe
-        >>> df.shape
-          (115047, 28)
-        ... # fetch data of 5 (randomly selected) stations
-        >>> df = dataset.fetch(stations=5, as_dataframe=True)
-        >>> df.shape
-           (115047, 5)
-        fetch data of 3 selected stations
-        >>> df = dataset.fetch(stations=['564','563','562'], as_dataframe=True)
-        >>> df.shape
-           (115047, 3)
-        ... # fetch data of a single stations
-        >>> df = dataset.fetch(stations='500', as_dataframe=True)
-           (115047, 1)
-        # get only selected dynamic features
-        >>> df = dataset.fetch(stations='501',
-        ...    dynamic_features=['AET_mm', 'Prec_mm',  'Streamflow_mm'], as_dataframe=True)
-        # fetch data between selected periods
-        >>> df = dataset.fetch(stations='225', st="20010101", en="20101231", as_dataframe=True)
-        >>> df.shape
-           (32868, 1)
-        ... # get data at monthly time step
-        >>> dataset = HYPE(time_step="month")
-        >>> df = dataset.fetch(stations='500', as_dataframe=True)
-        >>> df.shape
-           (3780, 1)
+    >>> from aqua_fetch import HYPE
+    >>> dataset = HYPE()
+    ... # get data by station id
+    >>> _, dynamic = dataset.fetch(stations='564', as_dataframe=True)
+    >>> df = dynamic['564'] # dynamic is a dictionary of with keys as station names and values as DataFrames
+    >>> df.shape
+    (12783, 9)
+
+    ... # get name of all stations as list
+    >>> stns = dataset.stations()
+    >>> len(stns)
+       564
+    ... # get data of 10 % of stations as dataframe
+    >>> _, dynamic = dataset.fetch(0.1, as_dataframe=True)
+    >>> len(dynamic)  # dynamic has data for 10% of stations (67 out of 671)
+       67
+
+    ... # dynamic is a dictionary whose values are dataframes of dynamic features
+    >>> [df.shape for df in dynamic.values()]
+        [(12783, 9), (12783, 9), (12783, 9),... (12783, 9), (12783, 9)]
+
+    ... get the data of a single (randomly selected) station
+    >>> _, dynamic = dataset.fetch(stations=1, as_dataframe=True)
+    >>> len(dynamic)  # dynamic has data for 1 station
+        1
+    ... # get names of available dynamic features
+    >>> dataset.dynamic_features
+    ... # get only selected dynamic features
+    >>> _, dynamic = dataset.fetch('564', as_dataframe=True,
+    ...  dynamic_features=['AET_mm', 'Prec_mm',  'Streamflow_mm'])
+    >>> dynamic['564'].shape
+       (12783, 3)
+
+    ... # get names of available static features
+    >>> dataset.static_features
+    ... # get data of 10 random stations
+    >>> _, dynamic = dataset.fetch(10, as_dataframe=True)
+    >>> len(dynamic)  # remember this is a dictionary with values as dataframe
+       10
+
+    # If we get both static and dynamic data
+    >>> static, dynamic = dataset.fetch(stations='564', static_features="all", as_dataframe=True)
+    >>> static.shape, len(dynamic), dynamic['564'].shape
+    ((1, 59), 1, (12783, 9))
+
[…]
