+++ aqua_fetch/rr/_brazil.py
@@ -55,42 +55,74 @@ class CAMELS_BR(_RainfallRunoff):
     --------
     >>> from aqua_fetch import CAMELS_BR
     >>> dataset = CAMELS_BR()
-    >>> _, df = dataset.fetch(stations=1, as_dataframe=True)
-    >>> df = df.unstack() # the returned dataframe is a multi-indexed dataframe so we have to unstack it
+    ... # get data by station id
+    >>> _, dynamic = dataset.fetch(stations='46035000', as_dataframe=True)
+    >>> df = dynamic['46035000'] # dynamic is a dictionary of with keys as station names and values as DataFrames
     >>> df.shape
-    (14245, 12)
-    # get name of all stations as list
+    (14245, 10)
+
+    ... # get name of all stations as list
     >>> stns = dataset.stations()
     >>> len(stns)
-    593
-    # we can get data of 10% catchments as below
-    >>> _, data = dataset.fetch(0.1, as_dataframe=True)
-    >>> data.shape
-    (170940, 59)
-    # the data is multi-index with ``time`` and ``dynamic_features`` as indices
-    >>> data.index.names == ['time', 'dynamic_features']
-     True
-    # get data by station id
-    >>> _, df = dataset.fetch(stations='46035000', as_dataframe=True)
-    >>> df.unstack().shape
-    (14245, 12)
-    # get names of available dynamic features
+       593
+    ... # get data of 10 % of stations as dataframe
+    >>> _, dynamic = dataset.fetch(0.1, as_dataframe=True)
+    >>> len(dynamic)  # dynamic has data for 10% of stations (59 out of 593)
+       59
+
+    ... # dynamic is a dictionary whose values are dataframes of dynamic features
+    >>> [df.shape for df in dynamic.values()]
+        [(14245, 10), (14245, 10), (14245, 10),... (14245, 10), (14245, 10)]
+
+    ... get the data of a single (randomly selected) station
+    >>> _, dynamic = dataset.fetch(stations=1, as_dataframe=True)
+    >>> len(dynamic)  # dynamic has data for 1 station
+        1
+    ... # get names of available dynamic features
     >>> dataset.dynamic_features
-    # get only selected dynamic features
-    >>> _, df = dataset.fetch(1, as_dataframe=True,
-    ... dynamic_features=['pcp_mm_cpc', 'aet_mm_mgb', 'airtemp_C_mean', 'q_cms_obs'])
-    >>> df.unstack().shape
-    (14245, 4)
-    # get names of available static features
+    ... # get only selected dynamic features
+    >>> _, dynamic = dataset.fetch('46035000', as_dataframe=True,
+    ...  dynamic_features=['pcp_mm_cpc', 'aet_mm_mgb', 'airtemp_C_mean', 'q_cms_obs'])
+    >>> dynamic['46035000'].shape
+       (14245, 4)
+
+    ... # get names of available static features
     >>> dataset.static_features
-    # get data of 10 random stations
-    >>> _, df = dataset.fetch(10, as_dataframe=True)
-    >>> df.shape
-    (170940, 10)  # remember this is multi-indexed DataFrame
-    # If we want to get both static and dynamic data
+    ... # get data of 10 random stations
+    >>> _, dynamic = dataset.fetch(10, as_dataframe=True)
+    >>> len(dynamic)  # remember this is a dictionary with values as dataframe
+       10
+
+    # If we get both static and dynamic data
     >>> static, dynamic = dataset.fetch(stations='46035000', static_features="all", as_dataframe=True)
-    >>> static.shape, dynamic.unstack().shape
-    ((1, 67), (14245, 12))
[…]
