### 2025-08-15 3fc334d
added dyn_fname dyn_fpath attributes of _RainfallRunoff class
