+++ aqua_fetch/rr/_usgs.py
@@ -61,7 +61,7 @@ class USGS(_RainfallRunoff):
     >>> df = dynamic['01010000'] # dynamic is a dictionary of with keys as station names and values as DataFrames
     >>> df.shape
     (27028, 20)
-
+    ...
     ... # get name of all stations as list
     >>> stns = dataset.stations()
     >>> len(stns)
@@ -70,11 +70,11 @@ class USGS(_RainfallRunoff):
     >>> _, dynamic = dataset.fetch(0.1, as_dataframe=True)
     >>> len(dynamic)  # dynamic has data for 10% of stations (1200 out of 12004)
        1200
-
+    ...
     ... # dynamic is a dictionary whose values are dataframes of dynamic features
     >>> [df.shape for df in dynamic.values()]
         [(27028, 20), (27028, 20), (27028, 20),... (27028, 20), (27028, 20)]
-
+    ...
     ... get the data of a single (randomly selected) station
     >>> _, dynamic = dataset.fetch(stations=1, as_dataframe=True)
     >>> len(dynamic)  # dynamic has data for 1 station
@@ -86,42 +86,42 @@ class USGS(_RainfallRunoff):
     ...  dynamic_features=['pcp_mm', 'snowmelt_mm', 'airtemp_C_2m_min', 'swe_mm', 'q_cms_obs'])
     >>> dynamic['01010000'].shape
        (27028, 4)
-
+    ...
     ... # get names of available static features
     >>> dataset.static_features
     ... # get data of 10 random stations
     >>> _, dynamic = dataset.fetch(10, as_dataframe=True)
     >>> len(dynamic)  # remember this is a dictionary with values as dataframe
        10
-
+    ...
     # If we get both static and dynamic data
     >>> static, dynamic = dataset.fetch(stations='01010000', static_features="all", as_dataframe=True)
     >>> static.shape, len(dynamic), dynamic['01010000'].shape
     ((1, 29), 1, (27028, 20))
-
+    ...
     # If we don't set as_dataframe=True and have xarray installed then the returned data will be a xarray Dataset
     >>> _, dynamic = dataset.fetch(10)
     ... type(dynamic)   
     xarray.core.dataset.Dataset
-
+    ...
     >>> dynamic.dims
     FrozenMappingWarningOnValuesAccess({'time': 27028, 'dynamic_features': 20})
-
+    ...
     >>> len(dynamic.data_vars)
     10
-
+    ...
     >>> coords = dataset.stn_coords() # returns coordinates of all stations
     >>> coords.shape
         (671, 2)
     >>> dataset.stn_coords('01010000')  # returns coordinates of station whose id is 01010000
         -69.715556	46.700556
     >>> dataset.stn_coords(['01010000', '01010070'])  # returns coordinates of two stations
-
+    ...
     # get area of a single station
     >>> dataset.area('01010000')
     # get coordinates of two stations
     >>> dataset.area(['01010000', '01010070'])
-
+    ...
     # if fiona library is installed we can get the boundary as fiona Geometry
     >>> dataset.get_boundary('01010000')
 
