### 2025-08-17 91cbe69
added verbosity arg to zonodo download
