+++ aqua_fetch/rr/_npctr.py
@@ -12,6 +12,7 @@ from ..utils import check_attributes
 from .utils import _RainfallRunoff
 from ._map import(
     observed_streamflow_cms,
+    observed_streamflow_mmd,
     mean_air_temp,
     mean_rel_hum,
     mean_windspeed,
@@ -110,6 +111,7 @@ class NPCTRCatchments(_RainfallRunoff):
 
     def __init__(self, path=None,
                  timestep:str = "Hourly",
+                 qflag = ["AV", "EV"],
                  **kwargs):
         super().__init__(path=path, **kwargs)
 
@@ -118,6 +120,18 @@ class NPCTRCatchments(_RainfallRunoff):
         self.timestep = _verify_timestep(timestep)
 
         self._static_features = self._get_static().columns.tolist()
+        
+        for flag in qflag:
+            assert flag in ['AV', 'EV', 'SVC', 'SVD']
+        self.qflags = qflag
+
+    @property
+    def start(self):
+        return pd.Timestamp("20130910 01:00:00")
+    
+    @property
+    def end(self):
+        return pd.Timestamp("20191231 23:00:00")
 
     @property
     def boundary_file(self) -> os.PathLike:
@@ -128,7 +142,16 @@ class NPCTRCatchments(_RainfallRunoff):
 
     @property
     def dynamic_features(self)->List[str]:
-        return []
+        return [total_precipitation(), 
+                observed_streamflow_cms(), 
+                'Qrate_min', 'Qrate_max', 'Qvol', 'Qvol_min',
+                'Qvol_max', 
+                observed_streamflow_mmd(), 
+       'Qmm_min', 'Qmm_max', 
+       mean_air_temp(),
+       mean_windspeed(), 
+       solar_radiation(), 
+       mean_rel_hum()]
 
     @property
     def static_features(self)->List[str]:
@@ -191,8 +214,19 @@ class NPCTRCatchments(_RainfallRunoff):
                              'Qrate': self.fp,
                              })
         df.index = pd.to_datetime(df.pop('Datetime'))
+        # drop the tz information from index
+        df.index = df.index.tz_localize(None)
+        df.index.name = 'time'
 
-        df.rename(columns={'Qrate': observed_streamflow_cms()}, inplace=True)
+        # drop rows where Qrate and Qvol are nans
+        df.dropna(subset=['Qrate', 'Qvol', 'Qmm'], how="all", inplace=True)
+
+        df.rename(columns={'Qrate': observed_streamflow_cms(),
+                           'Qmm': observed_streamflow_mmd()}, inplace=True)
+
+        df = df.loc[df['Qflag'].isin(self.qflags)]
+        # drop Qlevel and Qflag columns
+        df = df.drop(columns=['Qlevel', 'Qflag'])
 
         dfs = {name[3:]: grp for name, grp in df.groupby('Watershed')}
 
@@ -215,6 +249,27 @@ class NPCTRCatchments(_RainfallRunoff):
             dfs[stn] = df
         return dfs
 
+    def get_snowdepth(self):
+
+        data = self.read_snow_depth()
+
+        # todo : justify following selection
+        data = {
+            '1015': data['Hecate'],
+            '819': data['Hecate'],
+            '844': data['Hecate'],
+            '693': data['WSN693703'],
+            '703': data['WSN703708'],
+            '708': data['WSN703708'],
+            '626': data['WSN703708'],
+            }
+        
+        # drop duplicates from each dataframe based upon index
+        for stn, stn_df in data.items():
+            data[stn] = stn_df[~stn_df.index.duplicated(keep='first')]
+
+        return data
+
     def get_pcp(self, sensor="SSN"):
 
         df = self.read_pcp()
@@ -227,6 +282,15 @@ class NPCTRCatchments(_RainfallRunoff):
 
[…]
