+++ readme.md
@@ -195,7 +195,7 @@ mg_data_ohe.shape  # -> (1200, 31)
 | Japan          | 751                    | 696                     | 27                       | 35                      | 1979 - 2022       | Japan                                       | [Peirong et al., 2023](https://doi.org/10.5194/essd-16-1559-2024) & [river.go.jp](http://www1.river.go.jp)  |
 | LamaHCE        | 859                    | 859                     | 22                       | 80                      | 1981 - 2019       | Central Europe                              | [Klingler et al., 2021](https://doi.org/10.5194/essd-13-4529-2021)                                          |
 | LamaHIce       | 111                    | 111                     | 36                       | 154                     | 1950 - 2021       | Iceland                                     | [Helgason and Nijssen 2024](https://doi.org/10.5194/essd-16-2741-2024)                                      |
-| NPCTRCatchments| -                      | 7                       | 8                        | 14                      | 2013 - 2019       | Canada                                      | [Korver et al., 2022](https://doi.org/10.5194/essd-14-4231-2022)                                            |
+| NPCTRCatchments| -                      | 7                       | 14                       | 14                      | 2013 - 2019       | Canada                                      | [Korver et al., 2022](https://doi.org/10.5194/essd-14-4231-2022)                                            |
 | Poland         | 1287                   |                         | 10                       | 214                     | 1992 - 2020       | Poland                                      | [Nascimento et al., 2024](https://doi.org/10.5194/essd-2024-379) & [danepubliczne.imgw.pl](https://danepubliczne.imgw.pl) |
 | Portugal       | 280                    |                         | 10                       | 214                     | 1992 - 2020       | Portugal                                    | [Nascimento et al., 2024](https://doi.org/10.5194/essd-2024-379) & [SNIRH Portugal](https://snirh.apambiente.pt) |
 | RRLuleaSweden  | 1                      |                         | 2                        | 0                       | 2016 - 2019       | Lulea (Sweden)                              | [Broekhuizen et al., 2020](https://doi.org/10.5194/hess-24-869-2020)                                        |
