+++ paper/paper.md
@@ -36,7 +36,7 @@ The development of AquaFetch was inspired by the growing availability of diverse
 # Statement of need
 The rapid increase in publicly available spatio-temporal datasets in various domains, including water resources, is mainly driven by advances in computation and storage as well as demand to develop accurate data-driven solutions for challenges like climate change, water scarcity, and environmental pollution. However, importing them into Python is cumbersome, requiring researchers to navigate multiple sources and handle inconsistent formats and units. Many datasets also need extensive preprocessing, making its acquisition a complex and skill-intensive task before it can be used for analysis or modeling.
 
-These challenges highlight the need for a unified, consistent, automated, and reusable framework for extracting hydrological and environmental data. The AquaFetch package addresses this gap by leveraging data-handling tools such as Pandas [@Pandas], NumPy [@NumPy2020], xarray [@xarray2017], and netCDF5 [@NSF_Unidata_and_Davis_NetCDF-C] to offer a streamlined workflow for automatic data extraction from multiple sources in various formats.
+These challenges highlight the need for a unified, consistent, automated, and reusable framework for extracting hydrological and environmental data. The AquaFetch package addresses this gap by leveraging data-handling tools such as Pandas [@Pandas], NumPy [@NumPy2020], xarray [@xarray2017], and netCDF4 [@NSF_Unidata_and_Davis_NetCDF-C] to offer a streamlined workflow for automatic data extraction from multiple sources in various formats.
 
 Several other packages also aim to facilitate the acquisition and processing of water resource data. For example, tools like dataretrieval [@dataretrieval], hydrofunctions [@hydrofunctions], and harmonize-wq [@harmozie_wq] provide streamlined access to hydrological and water quality data, though they are primarily limited to data from U.S. Geological Survey (USGS) and Environmental Protection Agency (EPA) stations. Data Retriever [@retriever_Senyondo2017] offers access to over 200 parameters, including water quality, but lacks comprehensive rainfall-runoff datasets and wastewater treatment data. Meanwhile, tsp [@tsp] focuses on managing ground temperature data, and the HyRiver suite [@HyRiver] provides access to hydro-climate data exclusively from USGS’s National Water Information Service (NWIS). Geodata-Harvester [@GeodataHarvester] offers reusable and automated workflows for extracting diverse geospatial and environmental data for Australia.
 
@@ -60,7 +60,7 @@ Datasets are downloaded upon the first call to the respective function or class.
 ![Locations of measuring stations of water quality datasets\label{fig2}](wq_stations.png)
 
 # Testing and dependencies
-The package's core dependencies include requests [@requests], NumPy, and Pandas. xarray is utilized for saving data in netCDF5 [@NSF_Unidata_and_Davis_NetCDF-C] format, which is efficient for handling large datasets; however, this step is an optional dependency. Other optional dependencies include Matplotlib [@matplotlib] for plotting and visualization, openpyxl [@openpyxl] for parsing Microsoft Excel files, along with Fiona [@fiona] for processing shapefiles.
+The package's core dependencies include requests [@requests], NumPy, and Pandas. xarray is utilized for saving data in netCDF4 [@NSF_Unidata_and_Davis_NetCDF-C] format, which is efficient for handling large datasets; however, this step is an optional dependency. Other optional dependencies include Matplotlib [@matplotlib] for plotting and visualization, openpyxl [@openpyxl] for parsing Microsoft Excel files, along with Fiona [@fiona] for processing shapefiles.
 Adhering to the 'unit test' protocol, comprehensive testing has been implemented for all data classes and functions. Since downloading the datasets is time-consuming, the tests are conducted offline under the assumption that the datasets are already downloaded. These unit tests verify the number and types of parameters returned by the data functions.
 
 # Acknowledgements
