### 2025-08-21 97abf1f
resolved capitalization problem in .bib file
