+++ paper/paper.md
@@ -15,7 +15,7 @@ authors:
     orcid: 0000-0001-7446-6805
   - name: Hylke E. Beck
     corresponding: true
-    orcid: 0000-0002-0031-745X
+    orcid: 0000-0003-2553-9566
     affiliation: 1
 affiliations:
  - name: King Abdullah University of Science and Technology, Thuwal, Saudi Arabia
