+++ readme.md
@@ -168,7 +168,7 @@ mg_data_ohe.shape  # -> (1200, 31)
 | Arcticnet      | 106                    |                         | 27                       | 35                      | 1979 - 2003       | Arctic (Russia)                             | [R-Arcticnet](https://www.r-arcticnet.sr.unh.edu/v4.0/AllData/index.html)                                   |
 | Bull           | 484                    |                         | 55                       | 214                     | 1990 - 2020       | Spain                                       | [Aparicio et al., 2024](https://doi.org/10.1038/s41597-024-03594-5)                                         |
 | CABra          | 735                    |                         | 13                       | 87                      | 1980 - 2010       | Brazil                                      | [Almagro et al., 2021](https://doi.org/10.5194/hess-25-3105-2021)                                           |
-| CAMELSH        |                        | 5767                    | 13                       | 779                     | 1900 - 2024       | Unted Stated of America                     | [Tran et al., (2025)](https://doi.org/10.1038/s41597-025-05612-6)                                           |
+| CAMELSH        |                        | 5767                    | 13                       | 779                     | 1900 - 2024       | United States of America                     | [Tran et al., (2025)](https://doi.org/10.1038/s41597-025-05612-6)                                           |
 | CAMELS_AUS     | 222, 561               |                         | 28                       | 166, 187                | 1900 - 2018       | Australia                                   | [Flower et al., 2021](https://doi.org/10.5194/essd-13-3847-2021)                                            |
 | CAMELS_BR      | 897                    |                         | 10                       | 67                      | 1920 - 2019       | Brazil                                      | [Chagas et al., 2020](https://doi.org/10.5194/essd-12-2075-2020)                                            |
 | CAMELS_COL     | 347                    |                         | 6                        | 255                     | 1981 - 2022       | Columbia                                    | [Jimenez et al., 2025](https://doi.org/10.5194/essd-2025-200)                                               |
