### 2025-09-21 48dce27
updated CAMELS-DE download link to use latest dataset
