+++ setup.py
@@ -44,7 +44,7 @@ setup(
 
     name='aqua_fetch',
 
-    version = "1.0.0",
+    version = "1.0.1",
 
     description='A Unified Python Interface for Water Resource Data Acquisition and harmonization',
     long_description=long_desc,
