+++ aqua_fetch/__init__.py
@@ -280,4 +280,4 @@ def load_nasdaq(inputs: Union[str, list, None] = None, target: str = 'NDX'):
     return df[inputs + target]
 
 
-__version__ = "1.0.0"
+__version__ = "1.0.1"
