+++ readme.md
@@ -175,7 +175,7 @@ mg_data_ohe.shape  # -> (1200, 31)
 | CAMELS_CH      | 331                    |                         | 9                        | 209                     | 1981 - 2020       | Switzerland, Austria, France, Germany Italy | [Hoege et al., 2023](https://doi.org/10.5194/essd-15-5755-2023)                                             |
 | CAMELS_CL      | 516                    |                         | 12                       | 104                     | 1913 - 2018       | Chile                                       | [Alvarez-Garreton et al., 2018](https://doi.org/10.5194/hess-22-5817-2018)                                  |
 | CAMELS_DK      | 304                    |                         | 13                       | 119                     | 1989 - 2023       | Denmark                                     | [Liu et al., 2024](https://doi.org/10.5194/essd-2024-292)                                                   |
-| CAMELS_DE      | 1555                   |                         | 21                       | 111                     | 1951 - 2020       | Germany                                     | [Loritz et al., 2024](https://essd.copernicus.org/preprints/essd-2024-318)                                  |
+| CAMELS_DE      | 1582                   |                         | 21                       | 111                     | 1951 - 2020       | Germany                                     | [Loritz et al., 2024](https://doi.org/10.5194/essd-16-5625-2024)                                  |
 | CAMELS_FI      | 320                    |                         |                          | 111                     | 1963 - 2023       | Finland                                     | [Seppä, I et al., 2025](https://doi.org/10.5281/zenodo.16257216)                                            |
 | CAMELS_FR      | 654                    |                         | 22                       | 344                     | 1970 - 2021       | France                                      | [Delaigue et al., 2024](https://essd.copernicus.org/preprints/essd-2024-415/)                               |
 | CAMELS_GB      | 671                    |                         | 10                       | 145                     | 1970 - 2015       | Britain                                     | [Coxon et al., 2020](https://doi.org/10.5194/essd-12-2459-2020)                                             |
