### 2025-09-23 a87a6dd
updated download link for camels_us
